% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-activate_nebulae.R
\name{activate_nebulae-methods}
\alias{activate_nebulae-methods}
\alias{activate_nebulae}
\alias{activate_nebulae,missing,missing,missing-method}
\alias{activate_nebulae,mcnebula,ANY,ANY-method}
\alias{activate_nebulae,mcnebula,function,function-method}
\alias{ggset_activate_parent_nebula}
\alias{ggset_activate_child_nebulae}
\title{generate 'ggset' for visualization}
\usage{
\S4method{activate_nebulae}{missing,missing,missing}()

\S4method{activate_nebulae}{mcnebula,ANY,ANY}(x, fun_default_parent, fun_default_child)

\S4method{activate_nebulae}{mcnebula,`function`,`function`}(x, fun_default_parent, fun_default_child)

ggset_activate_parent_nebula(x)

ggset_activate_child_nebulae(x)
}
\arguments{
\item{x}{...}

\item{fun_default_parent}{...}

\item{fun_default_child}{...}
}
\value{
...

OUTPUT_DESCRIPTION

...
}
\description{
...

\code{activate_nebulae()}: get the default parameters for the method
\code{activate_nebulae}.

\code{activate_nebulae(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{activate_nebulae}.

...

FUNCTION_DESCRIPTION

...
}
\details{
...

...

DETAILS

...
}
\examples{
\dontrun{
activate_nebulae(...)
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
