% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-annotate_nebula.R
\name{annotate_nebula,ANY,character-method}
\alias{annotate_nebula,ANY,character-method}
\alias{annotate_nebula}
\title{...}
\usage{
\S4method{annotate_nebula}{ANY,character}(x, nebula_name)
}
\arguments{
\item{x}{...}

\item{nebula_name}{...}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
\dontrun{
annotate_nebula(...)
}
}
\seealso{
\code{\link{draw_nodes}}, \code{\link{draw_structures}}...
}
