% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-backtrack_stardust.R
\name{backtrack_stardust,mcnebula,missing,missing,missing-method}
\alias{backtrack_stardust,mcnebula,missing,missing,missing-method}
\alias{backtrack_stardust}
\alias{backtrack_stardust,mcnebula,character,missing,ANY-method}
\alias{backtrack_stardust,mcnebula,missing,numeric,ANY-method}
\title{...}
\usage{
\S4method{backtrack_stardust}{mcnebula,missing,missing,missing}(x)

\S4method{backtrack_stardust}{mcnebula,character,missing,ANY}(x, class.name, remove)

\S4method{backtrack_stardust}{mcnebula,missing,numeric,ANY}(x, rel.index, remove)
}
\arguments{
\item{x}{...}

\item{class.name}{...}

\item{remove}{...}

\item{rel.index}{...}
}
\value{
...
}
\description{
...

...

...
}
\details{
...
}
\examples{
\dontrun{
backtrack_stardust(...)
}
\dontrun{
backtrack_stardust(...)
}
\dontrun{
backtrack_stardust(...)
}
}
\seealso{
\code{\link{cross_filter_stardust}}
}
