% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraMethods-binary_comparison.R
\name{binary_comparison-methods}
\alias{binary_comparison-methods}
\alias{binary_comparison}
\alias{binary_comparison,missing,missing,missing,missing,missing-method}
\alias{binary_comparison,ANY,ANY,ANY,ANY,ANY-method}
\alias{binary_comparison,ANY,formula,function,ANY,character-method}
\title{...}
\usage{
\S4method{binary_comparison}{missing,missing,missing,missing,missing}()

\S4method{binary_comparison}{ANY,ANY,ANY,ANY,ANY}(x, ..., formula, fun_norm, top_coef, contrasts)

\S4method{binary_comparison}{ANY,formula,`function`,ANY,character}(x, formula, fun_norm, top_coef, contrasts)
}
\arguments{
\item{x}{...}

\item{...}{...}

\item{formula}{...}

\item{fun_norm}{...}

\item{top_coef}{...}

\item{contrasts}{...}
}
\value{
...
}
\description{
...

\code{binary_comparison()}: get the default parameters for the method
\code{binary_comparison}.

\code{binary_comparison(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{binary_comparison}.

...
}
\details{
...

...
}
\examples{
\dontrun{
binary_comparison(...)
}
}
\seealso{
\code{\link[stats:model.matrix]{stats::model.matrix()}}, \code{\link[limma:makeContrasts]{limma::makeContrasts()}}, \code{\link[limma:lmFit]{limma::lmFit()}},
\code{\link[limma:ebayes]{limma::eBayes()}}, \code{\link[limma:contrasts.fit]{limma::contrasts.fit()}}, \code{\link[limma:toptable]{limma::topTable()}}...
}
