% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-clear.R
\name{clear}
\alias{clear}
\alias{clear_dataset}
\alias{clear_nodes}
\title{Clean up data in the mcnebula object that is no longer in use}
\usage{
clear_dataset(x)

clear_nodes(x)
}
\arguments{
\item{x}{\linkS4class{mcnebula} object.}
}
\description{
\code{clear_dataset}: The data (chemical formula,
chemical structure, chemical classes)
in \code{project_dataset(x)}, and data in \code{backtrack(mcn_dataset(mcn)}
would be clean up to reduced memory usage.
This may be best used after running the \code{\link[=create_nebula_index]{create_nebula_index()}},
if your machine doesn't have much Random Access Memory (RAM).

\code{clear_nodes}: Clear data ('grobs' and 'ggset') of 'nodes'
in slot \code{child_nebulae}.
}
\note{
If this function has conducted, the PPCP dataset would not be
available for downstream methods, such as \code{\link[=set_ppcp_data]{set_ppcp_data()}},
\code{\link[=annotate_nebula]{annotate_nebula()}}...
}
