% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-section.R
\docType{class}
\name{code_block-class}
\alias{code_block-class}
\alias{.code_block}
\alias{code_block}
\alias{code_block_table-class}
\alias{.code_block_table}
\alias{code_block_table}
\alias{code_block_figure-class}
\alias{.code_block_figure}
\alias{code_block_figure}
\alias{show,code_block-method}
\alias{show}
\alias{show,code_block_table-method}
\alias{show,code_block_figure-method}
\alias{show,heading-method}
\alias{show,section-method}
\alias{code_block,ANY-method}
\alias{code_block<-,ANY-method}
\alias{code_block<-}
\alias{codes,code_block-method}
\alias{codes}
\alias{codes<-,code_block-method}
\alias{codes<-}
\alias{new_code_block,character,character,list,logical,function-method}
\alias{new_code_block}
\alias{new_code_block,missing,missing,missing,missing,missing-method}
\alias{new_code_block,ANY,ANY,ANY,ANY,ANY-method}
\alias{new_code_block_figure,character-method}
\alias{new_code_block_figure}
\alias{new_code_block_table,character-method}
\alias{new_code_block_table}
\alias{call_command,code_block-method}
\alias{call_command}
\title{...}
\usage{
\S4method{show}{code_block}(object)

\S4method{show}{code_block_table}(object)

\S4method{show}{code_block_figure}(object)

\S4method{show}{heading}(object)

\S4method{show}{section}(object)

\S4method{code_block}{ANY}(x)

\S4method{code_block}{ANY}(x) <- value

\S4method{codes}{code_block}(x)

\S4method{codes}{code_block}(x) <- value

\S4method{new_code_block}{character,character,list,logical,`function`}(language, codes, args, prettey, fun_prettey)

\S4method{new_code_block}{missing,missing,missing,missing,missing}()

\S4method{new_code_block}{ANY,ANY,ANY,ANY,ANY}(language, codes, args, prettey, fun_prettey)

\S4method{new_code_block_figure}{character}(name, caption, ...)

\S4method{new_code_block_table}{character}(name, ...)

\S4method{call_command}{code_block}(x)
}
\arguments{
\item{value}{The value for the slot.}

\item{language}{...}

\item{codes}{...}

\item{args}{...}

\item{prettey}{...}

\item{fun_prettey}{...}

\item{name}{...}

\item{caption}{...}

\item{...}{...}
}
\description{
...

...

...

\code{code_block}, \code{code_block<-}: getter and setter
for the \code{code_block} slot of the object.

\code{codes}, \code{codes<-}: getter and setter
for the \code{codes} slot of the object.

\code{new_code_block}: ...

\code{new_code_block()}: get the default parameters
for the method \code{new_code_block}.

\code{new_code_block(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{new_code_block}.

\code{new_code_block_figure}: ...

\code{new_code_block_table}: ...

\code{call_command}: Format 'code_block' object as character.
}
\section{Slots}{

\describe{
\item{\code{codes}}{...}

\item{\code{command_name}}{...}

\item{\code{command_function}}{...}

\item{\code{command_args}}{...}
}}

\examples{
\dontrun{
new('code_block', ...)
}
\dontrun{
new('code_block_table', ...)
}
\dontrun{
new('code_block_figure', ...)
}
\dontrun{
new_code_block(...)
}
\dontrun{
new_code_block_figure(...)
}
\dontrun{
new_code_block_table(...)
}
\dontrun{
call_command(...)
}
}
\seealso{
\code{\link{command-class}}

Other call_commands: 
\code{\link{ggset-class}},
\code{\link{report-class}},
\code{\link{section-class}}
}
\concept{call_commands}
