% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-command.R
\docType{class}
\name{command-class}
\alias{command-class}
\alias{.command}
\alias{command}
\alias{show,command-method}
\alias{show}
\alias{command_name,command-method}
\alias{command_name}
\alias{command_name<-,command-method}
\alias{command_name<-}
\alias{command_function,command-method}
\alias{command_function}
\alias{command_function<-,command-method}
\alias{command_function<-}
\alias{command_args,command-method}
\alias{command_args}
\alias{command_args<-,command-method}
\alias{command_args<-}
\alias{new_command,function,character-method}
\alias{new_command}
\alias{new_command,function,missing-method}
\alias{call_command,command-method}
\alias{call_command}
\title{...}
\usage{
\S4method{show}{command}(object)

\S4method{command_name}{command}(x)

\S4method{command_name}{command}(x) <- value

\S4method{command_function}{command}(x)

\S4method{command_function}{command}(x) <- value

\S4method{command_args}{command}(x)

\S4method{command_args}{command}(x) <- value

\S4method{new_command}{`function`,character}(fun, ..., name)

\S4method{new_command}{`function`,missing}(fun, ..., name)

\S4method{call_command}{command}(x)
}
\arguments{
\item{value}{The value for the slot.}

\item{fun}{...}

\item{...}{...}

\item{name}{...}
}
\description{
...

\code{command_name}, \code{command_name<-}: getter and setter
for the \code{command_name} slot of the object.

\code{command_function}, \code{command_function<-}: getter and setter
for the \code{command_function} slot of the object.

\code{command_args}, \code{command_args<-}: getter and setter
for the \code{command_args} slot of the object.

\code{new_command}: ...

\code{call_command}: ...
}
\section{Slots}{

\describe{
\item{\code{command_name}}{...}

\item{\code{command_function}}{...}

\item{\code{command_args}}{...}
}}

\examples{
\dontrun{
new('command', ...)
}
\dontrun{
new_command(...)
}
\dontrun{
call_command(...)
}
}
