% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-create_hierarchy.R
\name{create_hierarchy,mcnebula,function-method}
\alias{create_hierarchy,mcnebula,function-method}
\alias{create_hierarchy}
\alias{create_hierarchy,missing,missing-method}
\alias{create_hierarchy,mcnebula,ANY-method}
\title{...}
\usage{
\S4method{create_hierarchy}{mcnebula,`function`}(x, fun_organize)

\S4method{create_hierarchy}{missing,missing}()

\S4method{create_hierarchy}{mcnebula,ANY}(x, fun_organize)
}
\arguments{
\item{x}{...}

\item{fun_organize}{...}
}
\value{
...
}
\description{
...

\code{create_hierarchy()}: get the default parameters for the method
\code{create_hierarchy}.

\code{create_hierarchy(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{create_hierarchy}.
}
\details{
...
}
\examples{
\dontrun{
create_hierarchy(...)
}
}
