% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-create_parent_layout.R
\name{create_parent_layout,mcnebula,character,numeric-method}
\alias{create_parent_layout,mcnebula,character,numeric-method}
\alias{create_parent_layout}
\alias{create_parent_layout,missing,missing,missing-method}
\alias{create_parent_layout,mcnebula,ANY,ANY-method}
\title{...}
\usage{
\S4method{create_parent_layout}{mcnebula,character,numeric}(x, ggraph_layout, seed)

\S4method{create_parent_layout}{missing,missing,missing}()

\S4method{create_parent_layout}{mcnebula,ANY,ANY}(x, ggraph_layout, seed)
}
\arguments{
\item{x}{...}

\item{ggraph_layout}{...}

\item{seed}{...}
}
\value{
...
}
\description{
...

\code{create_parent_layout()}: get the default parameters for the method
\code{create_parent_layout}.

\code{create_parent_layout(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{create_parent_layout}.
}
\details{
...
}
\examples{
\dontrun{
create_parent_layout(...)
}
}
\seealso{
\code{\link[=fun]{fun()}}
}
