% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-create_stardust_classes.R
\name{create_stardust_classes,mcnebula,numeric,numeric,logical,logical-method}
\alias{create_stardust_classes,mcnebula,numeric,numeric,logical,logical-method}
\alias{create_stardust_classes}
\alias{create_stardust_classes,missing,missing,missing,missing,missing-method}
\alias{create_stardust_classes,mcnebula,ANY,ANY,ANY,ANY-method}
\title{...}
\usage{
\S4method{create_stardust_classes}{mcnebula,numeric,numeric,logical,logical}(
  x,
  pp.threashold,
  hierarchy_priority,
  position_isomerism,
  inherit_dataset
)

\S4method{create_stardust_classes}{missing,missing,missing,missing,missing}()

\S4method{create_stardust_classes}{mcnebula,ANY,ANY,ANY,ANY}(
  x,
  pp.threashold,
  hierarchy_priority,
  position_isomerism,
  inherit_dataset
)
}
\arguments{
\item{x}{...}

\item{pp.threashold}{...}

\item{hierarchy_priority}{...}

\item{position_isomerism}{...}

\item{inherit_dataset}{...}
}
\value{
...
}
\description{
...

\code{create_stardust_classes()}: get the default parameters for the method
\code{create_stardust_classes}.

\code{create_stardust_classes(x, ...)}:
use the default parameters whatever 'missing'
while performing the method \code{create_stardust_classes}.
}
\details{
...
}
\examples{
\dontrun{
create_stardust_classes(...)
}
}
