% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-xcms-feature_detection.R
\docType{class}
\name{detectFlow-class}
\alias{detectFlow-class}
\alias{.detectFlow}
\alias{detectFlow}
\alias{show_layers,detectFlow-method}
\alias{toBeEval-class}
\alias{toBeEval}
\alias{default_detectFlow}
\title{Steps in sequence to perform Feature Detection}
\usage{
\S4method{show_layers}{detectFlow}(x)

default_detectFlow(
  x,
  snthresh = 5,
  noise = 50000,
  peakwidth = c(3, 30),
  ppm = 20,
  minFraction = 0.1
)
}
\arguments{
\item{x}{\linkS4class{mcmass} object.}

\item{snthresh}{passed to \code{xcms::CentWaveParam} for
\code{xcms::findChromPeaks}}

\item{noise}{passed to \code{xcms::CentWaveParam} for
\code{xcms::findChromPeaks}}

\item{peakwidth}{passed to \code{xcms::CentWaveParam} for
\code{xcms::findChromPeaks}}

\item{ppm}{passed to \code{xcms::CentWaveParam} for
\code{xcms::findChromPeaks}}

\item{minFraction}{passed to \code{xcms::PeakDensityParam} for
\code{xcms::fillChromPeaks}}
}
\description{
A class inherits from \linkS4class{layerSet} to store steps for
Feature Detection.

See \code{show_layers} in \linkS4class{ggset}.

character(1) to be eval while perform \code{\link[=run_lcms]{run_lcms()}}.

\code{default_detectFlow}: Create \linkS4class{detectFlow} object for
Feature Detection.
}
\section{Slots}{

\describe{
\item{\code{layers}}{list. a list of objects \linkS4class{command}.}
}}

\examples{
\dontrun{
new('detectFlow', ...)
}
}
\seealso{
\linkS4class{layerSet}, \linkS4class{mcmass}
}
