% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraMethods-draw_nodes.R
\name{draw_nodes-methods}
\alias{draw_nodes-methods}
\alias{draw_nodes}
\alias{draw_nodes,missing,missing,missing,missing,missing,missing,missing-method}
\alias{draw_nodes,mcnebula,character,ANY,ANY,ANY,ANY,ANY-method}
\alias{draw_nodes,mcnebula,character,character,logical,logical,logical,logical-method}
\alias{show_node,missing,missing,missing,missing-method}
\alias{show_node,ANY,character,ANY,ANY-method}
\alias{show_node}
\alias{ggset_activate_nodes}
\title{...}
\usage{
\S4method{draw_nodes}{missing,missing,missing,missing,missing,missing,missing}()

\S4method{draw_nodes}{mcnebula,character,ANY,ANY,ANY,ANY,ANY}(
  x,
  nebula_name,
  nodes_color,
  add_id_text,
  add_structure,
  add_ppcp,
  add_ration
)

\S4method{draw_nodes}{mcnebula,character,character,logical,logical,logical,logical}(
  x,
  nebula_name,
  nodes_color,
  add_id_text,
  add_structure,
  add_ppcp,
  add_ration
)

\S4method{show_node}{missing,missing,missing,missing}()

\S4method{show_node}{ANY,character,ANY,ANY}(x, .features_id, panel_viewport, legend_viewport)

ggset_activate_nodes(
  x,
  .features_id,
  nodes_color = "#FFF9F2",
  add_ppcp = T,
  add_ration = T
)
}
\arguments{
\item{x}{...}

\item{nebula_name}{...}

\item{nodes_color}{...}

\item{add_id_text}{...}

\item{add_structure}{...}

\item{add_ppcp}{...}

\item{add_ration}{...}

\item{.features_id}{...}

\item{panel_viewport}{...}

\item{legend_viewport}{...}
}
\value{
...

...

OUTPUT_DESCRIPTION
}
\description{
...

\code{draw_nodes()}: get the function for generating
default parameters for the method
\code{draw_nodes}.

\code{draw_nodes(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{draw_nodes}.

...

\code{show_node()}: get the default parameters for the method
\code{show_node}.

...

\code{show_node(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{show_node}.

FUNCTION_DESCRIPTION
}
\details{
...

...

...

DETAILS
}
\examples{
\dontrun{
draw_nodes(...)
}
\dontrun{
show_node(...)
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
