% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-filter_formula.R
\name{filter_formula,mcnebula,function,logical-method}
\alias{filter_formula,mcnebula,function,logical-method}
\alias{filter_formula}
\alias{filter_formula,missing,missing,missing-method}
\alias{filter_formula,mcnebula,ANY,ANY-method}
\title{...}
\usage{
\S4method{filter_formula}{mcnebula,`function`,logical}(x, fun_filter, ..., by_reference)

\S4method{filter_formula}{missing,missing,missing}()

\S4method{filter_formula}{mcnebula,ANY,ANY}(x, fun_filter, ..., by_reference)
}
\arguments{
\item{param}{...}
}
\value{
...
}
\description{
...

\code{filter_formula()}: get the default parameters for the method
\code{filter_formula}.

\code{filter_formula(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{filter_formula}.
}
\details{
...
}
\examples{
\dontrun{
filter_formula(...)
}
}
\seealso{
\code{\link[=fun]{fun()}}
}
