% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-msframe.R
\name{filter_msframe,msframe,function,formula-method}
\alias{filter_msframe,msframe,function,formula-method}
\alias{filter_msframe}
\title{...}
\usage{
\S4method{filter_msframe}{msframe,`function`,formula}(x, fun_filter, f, ...)
}
\arguments{
\item{x}{...}

\item{fun_filter}{...}

\item{f}{...}

\item{...}{...}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
\dontrun{
filter_msframe(...)
}
}
\seealso{
\code{\link[=fun]{fun()}}
}
