% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-msframe.R
\name{format_msframe,msframe,missing,function,missing,function,missing-method}
\alias{format_msframe,msframe,missing,function,missing,function,missing-method}
\alias{format_msframe}
\title{...}
\usage{
\S4method{format_msframe}{msframe,missing,`function`,missing,`function`,missing}(x, fun_names, fun_types)
}
\arguments{
\item{x}{...}

\item{fun_names}{...}

\item{fun_types}{...}

\item{names}{...}

\item{types}{...}

\item{fun_format}{...}
}
\description{
...
}
\details{
...
}
\examples{
\dontrun{
format_msframe(...)
}
}
