% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ggset.R
\docType{class}
\name{ggset-class}
\alias{ggset-class}
\alias{.ggset}
\alias{ggset}
\alias{show_layers,ggset-method}
\alias{show_layers}
\alias{new_ggset,ANY-method}
\alias{new_ggset}
\alias{mutate_layer,ggset,numeric-method}
\alias{mutate_layer}
\alias{mutate_layer,ggset,character-method}
\alias{call_command,ggset-method}
\alias{call_command}
\title{Management for 'ggplot' visualzation}
\usage{
\S4method{show_layers}{ggset}(x)

\S4method{new_ggset}{ANY}(...)

\S4method{mutate_layer}{ggset,numeric}(x, layer, ...)

\S4method{mutate_layer}{ggset,character}(x, layer, ...)

\S4method{call_command}{ggset}(x)
}
\arguments{
\item{x}{\linkS4class{ggset} object}

\item{...}{parameters passed to the layer.}

\item{layer}{numeric(1) or character(1). If "character", the name must be unique
in slot \code{layers}.}
}
\description{
Let each packed "ggplot2" function (packed as \linkS4class{command} object)
into layers in sequence, allowing post modifications programmatically
and visualizing as "ggplot2" plot at any time.

\code{show_layers}: show functions and parameters in layers
with a pretty and readable form.

\code{new_ggset}: Simplified creation of \linkS4class{ggset} object.

\code{mutate_layer}:
Pass new parameters or modify pre-existing parameters to the packed function.

\code{call_command}: plot as 'ggplot' object.
}
\section{Slots}{

\describe{
\item{\code{layers}}{list with names. Each element of list must be a \linkS4class{command} object
packed 'ggplot2' function and its args.}
}}

\examples{
\dontrun{
show_layers(...)
}
\dontrun{
new_ggset(...)
}
\dontrun{
mutate_layer(...)
}
}
\seealso{
Other layerSets: 
\code{\link{layerSet-class}},
\code{\link{report-class}}

Other call_commands: 
\code{\link{code_block-class}},
\code{\link{command-class}},
\code{\link{report-class}},
\code{\link{section-class}}
}
\concept{call_commands}
\concept{layerSets}
