% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraMethods-report.R
\name{history_rblock-methods}
\alias{history_rblock-methods}
\alias{history_rblock}
\alias{history_rblock,missing,missing,missing,missing-method}
\alias{history_rblock,numeric,ANY,ANY,ANY-method}
\alias{history_rblock,numeric,missing,missing,numeric-method}
\alias{history_rblock,missing,character,character,ANY-method}
\title{Create 'code_block' object from history codes}
\usage{
\S4method{history_rblock}{missing,missing,missing,missing}()

\S4method{history_rblock}{numeric,ANY,ANY,ANY}(nrow, pattern_start, pattern_end, exclude)

\S4method{history_rblock}{numeric,missing,missing,numeric}(nrow, exclude)

\S4method{history_rblock}{missing,character,character,ANY}(pattern_start, pattern_end, exclude)
}
\arguments{
\item{nrow}{numeric(1). The number of lines of code to fetch.}

\item{pattern_start}{character(1).
The pattern string used to match the starting line of codes in R history.}

\item{pattern_end}{character(1).
The pattern string used to match the ending line of codes in R history.}

\item{exclude}{numeric(1). Used to exclude the last lines of code.}
}
\description{
Get codes from R history, then formatted as \linkS4class{code_block} object.

\code{history_rblock()}: get the default parameters for the method
\code{history_rblock}.

\code{history_rblock(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{history_rblock}.
}
\examples{
\dontrun{
  test1 <- 1
  test2 <- 2
  test3 <- 3
  
  block <- history_rblock(, "^test1", "^test3")
  block
}
}
\seealso{
\linkS4class{code_block}, \code{\link[=history]{history()}}...
}
