% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraMethods-report.R
\name{include_figure-methods}
\alias{include_figure-methods}
\alias{include_figure}
\alias{include_figure,character,character,character-method}
\title{Easily embed figure into document}
\usage{
\S4method{include_figure}{character,character,character}(file, name, caption)
}
\arguments{
\item{file}{character(1). The path of file. See \code{\link[knitr:include_graphics]{knitr::include_graphics()}} for
the supported image formats.}

\item{name}{character(1). For cross-reference in document.
See \url{https://bookdown.org/yihui/rmarkdown-cookbook/cross-ref.html#cross-ref}.}

\item{caption}{character(1). Caption of figure display in document.}
}
\description{
Creates a pre-defined \linkS4class{code_block_figure} object containing the codes of
\code{\link[knitr:include_graphics]{knitr::include_graphics()}} for formatting display the figure in document.
}
\examples{
\dontrun{
include_figure(...)
}
}
\seealso{
\linkS4class{code_block_figure}, \linkS4class{report}, \code{\link[knitr:include_graphics]{knitr::include_graphics()}}...
}
