% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.sirius.v6.R
\name{initialize_sirius_api}
\alias{initialize_sirius_api}
\title{Rsirius API tool}
\usage{
initialize_sirius_api(sirius, port = 8080L, ..., version = "v6", reset = FALSE)
}
\arguments{
\item{sirius}{character(1), the path to sirius.}

\item{port}{integer(1), port}

\item{version}{character(1), can ignore now}
}
\description{
The \code{initialize_sirius_api} used for connection with "http://localhost:{port}",
Then use \code{Rsirius} package to load data.
}
