% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-xcms-feature_detection.R
\docType{class}
\name{mcmass-class}
\alias{mcmass-class}
\alias{.mcmass}
\alias{mcmass}
\alias{raw_data,mcmass-method}
\alias{raw_data}
\alias{pro_data,mcmass-method}
\alias{pro_data}
\alias{pro_data<-,mcmass-method}
\alias{pro_data<-}
\alias{raw_data<-,mcmass-method}
\alias{raw_data<-}
\alias{sample_metadata,mcmass-method}
\alias{sample_metadata}
\alias{sample_metadata<-,mcmass-method}
\alias{sample_metadata<-}
\alias{features_quantification,mcmass-method}
\alias{features_quantification}
\alias{features_quantification<-,mcmass,data.frame-method}
\alias{features_quantification<-}
\alias{features_defination,mcmass-method}
\alias{features_defination}
\alias{features_defination<-,mcmass-method}
\alias{features_defination<-}
\alias{ms2_spectra,mcmass-method}
\alias{ms2_spectra}
\alias{ms2_spectra<-,mcmass-method}
\alias{ms2_spectra<-}
\alias{parameter_set,mcmass-method}
\alias{parameter_set}
\alias{parameter_set<-,mcmass,list-method}
\alias{parameter_set<-}
\alias{detectFlow,mcmass-method}
\alias{detectFlow}
\alias{detectFlow<-,mcmass,detectFlow-method}
\alias{detectFlow<-}
\alias{new_mcmass}
\title{storage of XCMS processed data of 'feature detection'}
\usage{
\S4method{raw_data}{mcmass}(x)

\S4method{pro_data}{mcmass}(x)

\S4method{pro_data}{mcmass}(x) <- value

\S4method{raw_data}{mcmass}(x) <- value

\S4method{sample_metadata}{mcmass}(x)

\S4method{sample_metadata}{mcmass}(x) <- value

\S4method{features_quantification}{mcmass}(x)

\S4method{features_quantification}{mcmass,data.frame}(x) <- value

\S4method{features_defination}{mcmass}(x)

\S4method{features_defination}{mcmass}(x) <- value

\S4method{ms2_spectra}{mcmass}(x)

\S4method{ms2_spectra}{mcmass}(x) <- value

\S4method{parameter_set}{mcmass}(x)

\S4method{parameter_set}{mcmass,list}(x) <- value

\S4method{detectFlow}{mcmass}(x)

\S4method{detectFlow}{mcmass,detectFlow}(x) <- value

new_mcmass(
  sample_metadata,
  snthresh = 5,
  noise = 50000,
  peakwidth = c(3, 30),
  ppm = 20,
  minFraction = 0.1
)
}
\arguments{
\item{value}{The value for the slot.}

\item{sample_metadata}{data.frame. Contains columns of 'file', 'sample',
'group'}

\item{snthresh}{passed to \code{xcms::CentWaveParam} for
\code{xcms::findChromPeaks}}

\item{noise}{passed to \code{xcms::CentWaveParam} for
\code{xcms::findChromPeaks}}

\item{peakwidth}{passed to \code{xcms::CentWaveParam} for
\code{xcms::findChromPeaks}}

\item{ppm}{passed to \code{xcms::CentWaveParam} for
\code{xcms::findChromPeaks}}

\item{minFraction}{passed to \code{xcms::PeakDensityParam} for
\code{xcms::fillChromPeaks}}
}
\description{
This class provides a process template for pre-processing
non-targeted mass spectrometry data with package \code{xcms}.
In default, the steps of Feature Detection were refer to:
\link{https://github.com/DorresteinLaboratory/XCMS3_FeatureBasedMN}

\code{raw_data}, \code{raw_data<-}: getter and setter
for the \code{raw_data} slot of the object.

\code{pro_data}, \code{pro_data<-}: getter and setter
for the \code{pro_data} slot of the object.

\code{features_quantification}, \code{features_quantification<-}: getter and setter
for the \code{features_quantification} slot of the object.

\code{features_defination}, \code{features_defination<-}: getter and setter
for the \code{features_defination} slot of the object.

\code{ms2_spectra}, \code{ms2_spectra<-}: getter and setter
for the \code{ms2_spectra} slot of the object.

\code{parameter_set}, \code{parameter_set<-}: getter and setter
for the \code{parameter_set} slot of the object.

\code{detectFlow}, \code{detectFlow<-}: getter and setter
for the \code{detectFlow} slot of the object.

\code{new_mcmass}: Create a \linkS4class{mcmass} object.
}
\section{Slots}{

\describe{
\item{\code{raw_data}}{\linkS4class{OnDiskMSnExp} object.}

\item{\code{pro_data}}{\linkS4class{XCMSnExp} object.}

\item{\code{sample_metadata}}{data.frame. User supplied data about the mass data
files to process.}

\item{\code{features_defination}}{\linkS4class{DFrame} object. Non user supplied data.
See \code{xcms::featureDefinitions}.}

\item{\code{features_quantification}}{data.frame. Non user supplied data.
Peak area were used to quantify the feature level.}

\item{\code{ms2_spectra}}{\linkS4class{MSpectra} object. Non user supplied data.
See \code{xcms::filteredMs2Spectra}.}

\item{\code{parameter_set}}{list. Passed to repalace the parameters in slot
\code{detectFlow} of \linkS4class{mcmass} for performing \code{\link[=run_lcms]{run_lcms()}}.}

\item{\code{detectFlow}}{\linkS4class{detectFlow} object.}
}}

\examples{
\dontrun{
new('mcmass', ...)
}
}
