% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mcn_dataset.R
\docType{class}
\name{mcn_dataset-class}
\alias{mcn_dataset-class}
\alias{.mcn_dataset}
\alias{mcn_dataset}
\alias{mcn_dataset,ANY-method}
\alias{mcn_dataset<-,ANY-method}
\alias{mcn_dataset<-}
\alias{latest,mcn_dataset,ANY,ANY-method}
\alias{latest}
\alias{extract_mcnset,ANY,character-method}
\alias{extract_mcnset}
\title{...}
\usage{
\S4method{mcn_dataset}{ANY}(x)

\S4method{mcn_dataset}{ANY}(x) <- value

\S4method{latest}{mcn_dataset,ANY,ANY}(x)

\S4method{extract_mcnset}{ANY,character}(x, subscript)
}
\arguments{
\item{x}{...}

\item{value}{The value for the slot.}

\item{subscript}{...}
}
\description{
...

\code{mcn_dataset}, \code{mcn_dataset<-}: getter and setter
for the \code{mcn_dataset} slot of the object.

\code{latest}: get the first data in \code{dataset} slot and
format as "tbl".

\code{extract_mcnset}: For fast extract data in object which containing
\code{mcn_dataset} slot. Normally not used.
}
\section{Slots}{

\describe{
\item{\code{dataset}}{...}

\item{\code{reference}}{...}

\item{\code{backtrack}}{...}
}}

\examples{
\dontrun{
new('mcn_dataset', ...)
}
\dontrun{
latest(...)
}
\dontrun{
extract_mcnset(...)
}
}
\seealso{
Other datasets: 
\code{\link{dataset-class}},
\code{\link{project_dataset-class}}

Other latests: 
\code{\link{mcnebula-class}},
\code{\link{msframe-class}},
\code{\link{project_dataset-class}}
}
\concept{datasets}
\concept{latests}
