% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mcnebula.R
\docType{class}
\name{mcnebula-class}
\alias{mcnebula-class}
\alias{.mcnebula}
\alias{mcnebula}
\alias{show,mcnebula-method}
\alias{show}
\alias{latest,mcnebula,character,ANY-method}
\alias{latest}
\alias{latest,missing,missing,missing-method}
\alias{latest,mcnebula,ANY,ANY-method}
\alias{creation_time,mcnebula-method}
\alias{creation_time}
\alias{creation_time<-,mcnebula-method}
\alias{creation_time<-}
\alias{ion_mode,mcnebula-method}
\alias{ion_mode}
\alias{ion_mode<-,mcnebula-method}
\alias{ion_mode<-}
\alias{palette_set,mcnebula-method}
\alias{palette_set}
\alias{palette_gradient,mcnebula-method}
\alias{palette_gradient}
\alias{palette_stat,mcnebula-method}
\alias{palette_stat}
\alias{palette_col,mcnebula-method}
\alias{palette_col}
\alias{palette_label,mcnebula-method}
\alias{palette_label}
\alias{reference,mcnebula-method}
\alias{reference}
\alias{specific_candidate,mcnebula-method}
\alias{specific_candidate}
\alias{hierarchy,mcnebula-method}
\alias{hierarchy}
\alias{stardust_classes,mcnebula-method}
\alias{stardust_classes}
\alias{nebula_index,mcnebula-method}
\alias{nebula_index}
\alias{spectral_similarity,mcnebula-method}
\alias{spectral_similarity}
\alias{spectral_similarity<-,mcnebula-method}
\alias{spectral_similarity<-}
\alias{features_annotation,mcnebula-method}
\alias{features_annotation}
\alias{features_quantification,mcnebula-method}
\alias{features_quantification}
\alias{features_quantification<-,mcnebula-method}
\alias{sample_metadata,mcnebula-method}
\alias{sample_metadata}
\alias{sample_metadata<-,mcnebula-method}
\alias{sample_metadata<-}
\alias{classification,mcnebula-method}
\alias{classification}
\title{...}
\usage{
\S4method{show}{mcnebula}(object)

\S4method{latest}{mcnebula,character,ANY}(x, slot, subscript)

\S4method{latest}{missing,missing,missing}()

\S4method{latest}{mcnebula,ANY,ANY}(x, slot, subscript)

\S4method{creation_time}{mcnebula}(x)

\S4method{creation_time}{mcnebula}(x) <- value

\S4method{ion_mode}{mcnebula}(x)

\S4method{ion_mode}{mcnebula}(x) <- value

\S4method{palette_set}{mcnebula}(x)

\S4method{palette_gradient}{mcnebula}(x)

\S4method{palette_stat}{mcnebula}(x)

\S4method{palette_col}{mcnebula}(x)

\S4method{palette_label}{mcnebula}(x)

\S4method{reference}{mcnebula}(x)

\S4method{specific_candidate}{mcnebula}(x)

\S4method{hierarchy}{mcnebula}(x)

\S4method{stardust_classes}{mcnebula}(x)

\S4method{nebula_index}{mcnebula}(x)

\S4method{spectral_similarity}{mcnebula}(x)

\S4method{spectral_similarity}{mcnebula}(x) <- value

\S4method{features_annotation}{mcnebula}(x)

\S4method{features_quantification}{mcnebula}(x)

\S4method{features_quantification}{mcnebula}(x) <- value

\S4method{sample_metadata}{mcnebula}(x)

\S4method{sample_metadata}{mcnebula}(x) <- value

\S4method{classification}{mcnebula}(x)
}
\arguments{
\item{x}{\linkS4class{mcnebula} object}

\item{slot}{Character. Slot name.}

\item{subscript}{numeric or character. The sequence or name for dataset in the 'list'.}

\item{value}{The value for the slot.}
}
\description{
...

\code{latest(x, slot, subscript)}: get the data in slot
(\code{mcn_dataset(object)} or \code{prject_dataset(object)})
and format as 'tbl'.

\code{latest()}: get the default parameters for the method \code{latest}.

\code{latest(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{latest}.

\code{creation_time}, \code{creation_time<-}: getter and setter
for the \code{creation_time} slot of the object.

\code{ion_mode}, \code{ion_mode<-}: getter and setter
for the \code{ion_mode} slot of the object.

\code{palette_set}, \code{palette_gradient}, \code{palette_stat},
\code{palette_col}: fast channel to obtain the downstream slot.
For \code{palette_set}, e.g., getter for the \code{palette_set} slot in sub-object
of \code{melody} slot of the object. \code{palette_set(melody(object))}
equals to \code{palette_set(object)}.

\code{reference}: fast channel to obtain
the downstream slot, getter
for the \code{reference} slot in sub-object
of \code{mcn_dataset} slot of the object.
\code{reference(mcn_dataset(object))}
equals to \code{reference(object)}.

\code{specific_candidate}, \code{hierarchy}, \code{stardust_classes},
\code{nebula_index}, \code{spectral_similarity}, \code{features_annotation},
\code{features_quantification}, \code{sample_metadata}:
fast channel to obtain data (mostly 'tbl' or 'data.frame')
inside the downstream slot ('list'). e.g., getter
for the data named \code{specific_candidate} in
\code{reference} slot (a 'list') in sub-object
of \code{mcn_dataset} slot of the object.
\code{reference(mcn_dataset(object))$specific_candidate}
equals to \code{specific_candidate(object)}.

\code{spectral_similarity<-}, \code{features_quantification<-},
\code{sample_metadata<-}: fast channel to replace
data (mostly 'tbl' or 'data.frame') inside the downstream slot ('list'). e.g., setter
for the data named \code{spectral_similarity} in
\code{reference} slot (a 'list') in sub-object
of \code{mcn_dataset} slot of the object.
\code{reference(mcn_dataset(object))$spectral_similarity<-}
similar to \code{spectral_similarity(object)<-}.
But the latter not only replace and also validate.

\code{classification}: fast channel to obtain
data deeply inside the downstream slot ('list'), getter
for the data named \code{".canopus"} in
\code{dataset} slot (a 'list') in sub-object
of \code{project_dataset} slot of the object.
\code{tibble::as_tibble(entity(dataset(project_dataset(object))$.canopus))}
equals to \code{classification(object)}.
}
\section{Slots}{

\describe{
\item{\code{creation_time}}{...}

\item{\code{ion_mode}}{...}

\item{\code{melody}}{...}

\item{\code{mcn_dataset}}{...}

\item{\code{statistic_set}}{...}

\item{\code{project_version}}{...}

\item{\code{project_path}}{...}

\item{\code{project_conformation}}{...}

\item{\code{project_metadata}}{...}

\item{\code{project_api}}{...}

\item{\code{project_dataset}}{...}

\item{\code{parent_nebula}}{...}

\item{\code{child_nebulae}}{...}

\item{\code{export_path}}{...}

\item{\code{export_name}}{...}
}}

\examples{
\dontrun{
new('mcnebula', ...)
}
\dontrun{
latest(x)
latest(x, "project_dataset")
latest(x, "mcn_dataset")
}
}
\seealso{
\code{\link[=fun]{fun()}}

\code{\link[tibble:as_tibble]{tibble::as_tibble()}}

Other nebulae: 
\code{\link{parent_nebula-class}}

Other latests: 
\code{\link{mcn_dataset-class}},
\code{\link{msframe-class}},
\code{\link{project_dataset-class}}

Other subscripts: 
\code{\link{msframe-class}},
\code{\link{subscript-class}}
}
\concept{latests}
\concept{nebulae}
\concept{subscripts}
