% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-project_dataset.R
\docType{class}
\name{project_dataset-class}
\alias{project_dataset-class}
\alias{.project_dataset}
\alias{project_dataset}
\alias{project_dataset,ANY-method}
\alias{project_dataset<-,ANY-method}
\alias{project_dataset<-}
\alias{latest,project_dataset,ANY,ANY-method}
\alias{latest}
\alias{extract_rawset,ANY,character,ANY-method}
\alias{extract_rawset}
\alias{extract_rawset,ANY,character,function-method}
\title{Store extracted data}
\usage{
\S4method{project_dataset}{ANY}(x)

\S4method{project_dataset}{ANY}(x) <- value

\S4method{latest}{project_dataset,ANY,ANY}(x)

\S4method{extract_rawset}{ANY,character,ANY}(x, subscript)

\S4method{extract_rawset}{ANY,character,`function`}(x, subscript, fun_collate, ...)
}
\arguments{
\item{x}{an object contain \code{project_dataset} slot.}

\item{value}{The value for the slot.}

\item{subscript}{character. Specified the data in \code{dataset} slot
in \code{project_dataset} slot.
See \code{\link{VIRTUAL_subscript-class}}.}

\item{fun_collate}{function. If the specified data not exists in \code{dataset} slot,
it will be used to collate data. This parameter is not for normal use.}

\item{...}{parameters passed to 'fun_collate'.}
}
\description{
This is a class object used to store extracted data (raw data).
See \linkS4class{project} for joint application with other related classes.

\code{project_dataset}, \code{project_dataset<-}: getter and setter
for the \code{project_dataset} slot of the object.

\code{latest}: get the first data in \code{dataset} slot ('list') and
format as 'tbl'. Equals:
\itemize{
\item \code{latest(object)}
\item \code{tibble::as_tibble(entity(dataset(object)[[1]]))}
}

\code{extract_rawset}: For fast extract data in object which containing
\code{project_dataset} slot. Normally not used.
}
\section{Slots}{

\describe{
\item{\code{dataset}}{list. See \linkS4class{dataset}.}
}}

\examples{
\dontrun{
new('project_dataset', ...)
}
\dontrun{
latest(object)
}
\dontrun{
extract_rawset(object, ".f3_fingerid")
}
}
\seealso{
\code{\link[tibble:as_tibble]{tibble::as_tibble()}}

Other projects: 
\code{\link{project-class}},
\code{\link{project_api-class}},
\code{\link{project_conformation-class}},
\code{\link{project_metadata-class}}

Other datasets: 
\code{\link{dataset-class}},
\code{\link{mcn_dataset-class}}

Other latests: 
\code{\link{mcn_dataset-class}},
\code{\link{mcnebula-class}},
\code{\link{msframe-class}},
\code{\link{project_metadata-class}}
}
\concept{datasets}
\concept{latests}
\concept{projects}
