% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-project_metadata.R
\docType{class}
\name{project_metadata-class}
\alias{project_metadata-class}
\alias{.project_metadata}
\alias{project_metadata}
\alias{show,project_metadata-method}
\alias{show}
\alias{project_metadata,ANY-method}
\alias{project_metadata<-,ANY-method}
\alias{project_metadata<-}
\alias{latest,project_metadata,ANY,ANY-method}
\alias{latest}
\alias{metadata,project_metadata-method}
\alias{metadata}
\alias{metadata<-,project_metadata-method}
\alias{metadata<-}
\alias{add_dataset,project_metadata,list-method}
\alias{add_dataset}
\alias{extract_metadata,ANY,character-method}
\alias{extract_metadata}
\alias{get_metadata,ANY,character,ANY,ANY,ANY-method}
\alias{get_metadata}
\alias{get_metadata,missing,character,project_metadata,project_conformation,character-method}
\title{...}
\usage{
\S4method{show}{project_metadata}(object)

\S4method{project_metadata}{ANY}(x)

\S4method{project_metadata}{ANY}(x) <- value

\S4method{latest}{project_metadata,ANY,ANY}(x)

\S4method{metadata}{project_metadata}(x)

\S4method{metadata}{project_metadata}(x) <- value

\S4method{add_dataset}{project_metadata,list}(x, list)

\S4method{extract_metadata}{ANY,character}(x, subscript)

\S4method{get_metadata}{ANY,character,ANY,ANY,ANY}(x, subscript)

\S4method{get_metadata}{missing,character,project_metadata,project_conformation,character}(subscript, project_metadata, project_conformation, path)
}
\arguments{
\item{x}{...}

\item{value}{The value for the slot.}

\item{list}{...}

\item{subscript}{...}

\item{project_metadata}{...}

\item{project_conformation}{...}

\item{path}{...}
}
\description{
...

\code{project_metadata}, \code{project_metadata<-}: getter and setter
for the \code{project_metadata} slot of the object.

\code{latest}: get the first data in \code{metadata} slot and
format as "tbl".

\code{metadata}, \code{metadata<-}: getter and setter
for the \code{metadata} slot of the object.

\code{add_dataset}: ...

\code{extract_metadata}: ...

\code{get_metadata}: ...
}
\section{Slots}{

\describe{
\item{\code{metadata}}{...}
}}

\examples{
\dontrun{
new('project_metadata', ...)
}
\dontrun{
latest(...)
}
\dontrun{
add_dataset(...)
}
\dontrun{
extract_metadata(...)
}
\dontrun{
get_metadata(...)
}
}
\seealso{
Other projects: 
\code{\link{project-class}},
\code{\link{project_api-class}},
\code{\link{project_conformation-class}},
\code{\link{project_dataset-class}}
}
\concept{projects}
