% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraMethods-report.R
\name{rblock}
\alias{rblock}
\title{Eval the code as well create 'code_block' object}
\usage{
rblock(code, eval = T, envir = parent.frame())
}
\arguments{
\item{code}{The code to run or document. Braces (‘{}’) must be used.}

\item{eval}{logical. Whether to eval the code.}

\item{envir}{environment. The ‘environment’ in which the code is to be evaluated.}
}
\description{
\code{rblock}: Run or not run the code with formatting as \linkS4class{code_block}
object.
}
\examples{
\dontrun{
  rblock({
    test1 <- 1
    test2 <- 2
    test3 <- 3
  })
  
  rblock({
    test <- mcn_5features
    ## this annotation line would be ignored
    test1 <- filter_structure(test)
    test1 <- create_reference(test1)
    test1 <- filter_formula(test1, by_reference=T)
    test1 <- create_stardust_classes(test1)
  })
}
}
