% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-report.R
\name{render_report}
\alias{render_report}
\title{Convert 'report' as .Rmd file}
\usage{
render_report(x, file, set_all_eval = F)
}
\arguments{
\item{x}{\linkS4class{report} object.}

\item{file}{character(1). File name to save as.}

\item{set_all_eval}{logical(1). If \code{TRUE}, all \linkS4class{code_block} object
or \linkS4class{section} object related with "r" would be set with \code{eval = T}.}
}
\description{
\code{render_report}: Write down \linkS4class{report} onject
as .Rmd file, then \code{\link[rmarkdown:render]{rmarkdown::render()}} can be used to output any
available formation.
}
