% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reportDoc}
\alias{reportDoc}
\title{Example text for report description.}
\format{
\subsection{\code{reportDoc}}{

A list object.
}
}
\usage{
reportDoc
}
\description{
Lazy data used to supplement the presentation of the report.
It doesn't make the description of the report outstanding, but it at least
makes it decent (maybe).
}
\keyword{datasets}
