% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-xcms-feature_detection.R
\name{run_lcms,mcmass-method}
\alias{run_lcms,mcmass-method}
\alias{run_lcms}
\alias{run_export}
\title{Perform Feature Detection with steps in 'detectFlow'}
\usage{
\S4method{run_lcms}{mcmass}(x)

run_export(x, keep_onlyWithMs2 = T, mzd = 0, minProp = 0.3, ppm = 20, ...)
}
\arguments{
\item{x}{\linkS4class{mcmass} object.}

\item{keep_onlyWithMs2}{logical(1). If \code{TRUE}, the data
\code{features_quantification} only keep features which possess MS2.}

\item{mzd}{passed to \code{MSnbase::combineSpectra}}

\item{minProp}{passed to \code{MSnbase::combineSpectra}}

\item{ppm}{passed to \code{MSnbase::combineSpectra}}

\item{...}{passed to \code{MSnbase::combineSpectra}}
}
\description{
This would use package mainly 'xcms' to perform
Feature Detection.

\code{run_export}: get \code{features_quantification}
and \code{ms2_spectra} in \linkS4class{mcmass} object.
}
\examples{
\dontrun{
run_lcms(...)
}
}
