% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-section.R
\docType{class}
\name{heading-class}
\alias{heading-class}
\alias{.heading}
\alias{heading}
\alias{section-class}
\alias{.section}
\alias{section}
\alias{heading,ANY-method}
\alias{heading<-,ANY-method}
\alias{heading<-}
\alias{level,heading-method}
\alias{level}
\alias{level<-,heading-method}
\alias{level<-}
\alias{new_heading,character,numeric-method}
\alias{new_heading}
\alias{call_command,heading-method}
\alias{call_command}
\alias{paragraph,section-method}
\alias{paragraph}
\alias{paragraph<-,section-method}
\alias{paragraph<-}
\alias{new_section,missing,missing,missing,missing-method}
\alias{new_section,ANY,ANY,ANY,ANY-method}
\alias{new_section,character,numeric,character,maybe_code_block-method}
\alias{new_section}
\alias{new_section,NULL,numeric,character,maybe_code_block-method}
\alias{call_command,section-method}
\alias{call_command,NULL-method}
\title{...}
\usage{
\S4method{heading}{ANY}(x)

\S4method{heading}{ANY}(x) <- value

\S4method{level}{heading}(x)

\S4method{level}{heading}(x) <- value

\S4method{new_heading}{character,numeric}(heading, level)

\S4method{call_command}{heading}(x)

\S4method{paragraph}{section}(x)

\S4method{paragraph}{section}(x) <- value

\S4method{new_section}{missing,missing,missing,missing}(heading, level, paragraph, code_block)

\S4method{new_section}{ANY,ANY,ANY,ANY}(heading, level, paragraph, code_block)

\S4method{new_section}{character,numeric,character,maybe_code_block}(heading, level, paragraph, code_block)

\S4method{new_section}{`NULL`,numeric,character,maybe_code_block}(heading, level, paragraph, code_block)

\S4method{call_command}{section}(x)

\S4method{call_command}{`NULL`}(x)
}
\arguments{
\item{value}{The value for the slot.}

\item{heading}{...}

\item{level}{...}

\item{paragraph}{...}

\item{code_block}{...}
}
\description{
...

...

\code{heading}, \code{heading<-}: getter and setter
for the \code{heading} slot of the object.

\code{level}, \code{level<-}: getter and setter
for the \code{level} slot of the object.

\code{new_heading}: ...

\code{call_command}: Format 'heading' object as character.

\code{paragraph}, \code{paragraph<-}: getter and setter
for the \code{paragraph} slot of the object.

\code{new_section()}: get the default parameters for
the method \code{new_section}.

\code{new_section(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{new_section}.

\code{new_section}: ...

\code{call_command}: Format 'section' object as character.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{...}

\item{\code{level}}{...}

\item{\code{heading}}{...}

\item{\code{paragraph}}{...}

\item{\code{code_block}}{...}
}}

\examples{
\dontrun{
new('heading', ...)
}
\dontrun{
new('section', ...)
}
\dontrun{
new_heading(...)
}
\dontrun{
call_command(...)
}
\dontrun{
new_section(...)
}
\dontrun{
call_command(...)
}
}
\seealso{
Other call_commands: 
\code{\link{code_block-class}},
\code{\link{report-class}}
}
\concept{call_commands}
