% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-create_nebula_index.R
\name{set_tracer-methods}
\alias{set_tracer-methods}
\alias{set_tracer}
\alias{set_tracer,missing,missing,missing,missing-method}
\alias{set_tracer,mcnebula,ANY,ANY,ANY-method}
\alias{set_tracer,mcnebula,character,character,character-method}
\title{Mark top 'features' in 'nebula_index' data}
\usage{
\S4method{set_tracer}{missing,missing,missing,missing}()

\S4method{set_tracer}{mcnebula,ANY,ANY,ANY}(x, .features_id, colors, rest)

\S4method{set_tracer}{mcnebula,character,character,character}(x, .features_id, colors, rest)
}
\arguments{
\item{x}{\linkS4class{mcnebula} object.}

\item{.features_id}{character. The ID of 'features' to mark.}

\item{colors}{character. Hex color.}

\item{rest}{character(1). Hex color.}
}
\description{
Custom defined the specific 'features' in 'nebula_index' data.
Mark these 'features' for subsequent visualization with eye-catching highlighting
(\code{\link[=set_nodes_color]{set_nodes_color()}}).
Run after \code{\link[=create_nebula_index]{create_nebula_index()}}.

\code{set_tracer()}: get the function for generating
default parameters for the method
\code{set_tracer}.

\code{set_tracer(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{set_tracer}.
}
\examples{
\dontrun{
  test <- mcn_5features
  
  ## the previous steps
  test1 <- filter_structure(test)
  test1 <- create_reference(test1)
  test1 <- filter_formula(test1, by_reference = T)
  test1 <- create_stardust_classes(test1)
  test1 <- create_features_annotation(test1)
  test1 <- cross_filter_stardust(test1, 2, 1)
  test1 <- create_nebula_index(test1)
  
  ids <- features_annotation(test1)$.features_id
  test1 <- set_tracer(test1, ids[1:2])
  ## see results
  nebula_index(test1)
  
  ## see examples in 'set_nodes_color()'
}
}
\seealso{
\code{\link[=create_nebula_index]{create_nebula_index()}}, \code{\link[=set_nodes_color]{set_nodes_color()}}.
}
