% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-statistic_set.R
\docType{class}
\name{statistic_set-class}
\alias{statistic_set-class}
\alias{.statistic_set}
\alias{statistic_set}
\alias{statistic_set,ANY-method}
\alias{statistic_set<-,ANY-method}
\alias{statistic_set<-}
\alias{design_matrix,ANY-method}
\alias{design_matrix}
\alias{design_matrix<-,ANY-method}
\alias{design_matrix<-}
\alias{contrast_matrix,ANY-method}
\alias{contrast_matrix}
\alias{contrast_matrix<-,ANY-method}
\alias{contrast_matrix<-}
\alias{top_table,ANY-method}
\alias{top_table}
\alias{top_table<-,ANY-method}
\alias{top_table<-}
\title{...}
\usage{
\S4method{statistic_set}{ANY}(x)

\S4method{statistic_set}{ANY}(x) <- value

\S4method{design_matrix}{ANY}(x)

\S4method{design_matrix}{ANY}(x) <- value

\S4method{contrast_matrix}{ANY}(x)

\S4method{contrast_matrix}{ANY}(x) <- value

\S4method{top_table}{ANY}(x)

\S4method{top_table}{ANY}(x) <- value
}
\arguments{
\item{value}{The value for the slot.}
}
\description{
...

\code{statistic_set}, \code{statistic_set<-}: getter and setter
for the \code{statistic_set} slot of the object.

\code{design_matrix}, \code{design_matrix<-}: getter and setter
for the \code{design_matrix} slot of the object.

\code{contrast_matrix}, \code{contrast_matrix<-}: getter and setter
for the \code{contrast_matrix} slot of the object.

\code{top_table}, \code{top_table<-}: getter and setter
for the \code{top_table} slot of the object.
}
\section{Slots}{

\describe{
\item{\code{design_matrix}}{...}

\item{\code{contrast_matrix}}{...}

\item{\code{dataset}}{...}

\item{\code{top_table}}{...}
}}

\examples{
\dontrun{
new('statistic_set', ...)
}
}
