% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VIRTUAL_slots.R
\docType{class}
\name{dataset-class}
\alias{dataset-class}
\alias{VIRTUAL_dataset}
\alias{dataset}
\alias{dataset,ANY-method}
\alias{dataset<-,ANY-method}
\alias{dataset<-}
\title{Share slots and methods for classes inherite from VIRTUAL_dataset}
\usage{
\S4method{dataset}{ANY}(x)

\S4method{dataset}{ANY}(x) <- value
}
\arguments{
\item{value}{The value for the slot.}
}
\description{
...

\code{dataset}, \code{dataset<-}: getter and setter
for the \code{dataset} slot of the object.
}
\section{Slots}{

\describe{
\item{\code{dataset}}{...}
}}

\seealso{
Other datasets: 
\code{\link{mcn_dataset-class}},
\code{\link{project_dataset-class}}
}
\concept{datasets}
