% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VIRTUAL_slots.R
\docType{class}
\name{export-class}
\alias{export-class}
\alias{VIRTUAL_export}
\alias{export}
\alias{export_name,ANY-method}
\alias{export_name}
\alias{export_name<-,ANY-method}
\alias{export_name<-}
\alias{export_path,ANY-method}
\alias{export_path}
\alias{export_path<-,ANY-method}
\alias{export_path<-}
\title{Share slots and methods for classes inherite from VIRTUAL_export}
\usage{
\S4method{export_name}{ANY}(x)

\S4method{export_name}{ANY}(x) <- value

\S4method{export_path}{ANY}(x)

\S4method{export_path}{ANY}(x) <- value
}
\arguments{
\item{value}{The value for the slot.}
}
\description{
This VIRTUAL class provides slots for recording export path
and export name of attributes.

\code{export_name}, \code{export_name<-}: getter and setter
for the \code{export_name} slot of the object.

\code{export_path}, \code{export_path<-}: getter and setter
for the \code{export_path} slot of the object.
}
\section{Slots}{

\describe{
\item{\code{export_path}}{character(1). The export directory path.}

\item{\code{export_name}}{character with names.
While export, the attribute name will be converted to the value.}
}}

\concept{exports}
