% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VIRTUAL_slots.R
\docType{class}
\name{reference-class}
\alias{reference-class}
\alias{VIRTUAL_reference}
\alias{reference}
\alias{reference,ANY-method}
\alias{reference<-,ANY-method}
\alias{reference<-}
\title{Share slots and methods for classes inherite from VIRTUAL_reference}
\usage{
\S4method{reference}{ANY}(x)

\S4method{reference}{ANY}(x) <- value
}
\arguments{
\item{value}{The value for the slot.}
}
\description{
...

\code{reference}, \code{reference<-}: getter and setter
for the \code{reference} slot of the object.
}
\section{Slots}{

\describe{
\item{\code{reference}}{...}
}}

\concept{references}
