% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VIRTUAL_slots.R
\docType{class}
\name{subscript-class}
\alias{subscript-class}
\alias{VIRTUAL_subscript}
\alias{subscript}
\alias{subscript,ANY-method}
\alias{subscript<-,ANY-method}
\alias{subscript<-}
\title{Share slots and methods for classes inherite from VIRTUAL_subscript}
\usage{
\S4method{subscript}{ANY}(x)

\S4method{subscript}{ANY}(x) <- value
}
\arguments{
\item{value}{The value for the slot.}
}
\description{
...

\code{subscript}, \code{subscript<-}: getter and setter
for the \code{subscript} slot of the object.
}
\section{Slots}{

\describe{
\item{\code{subscript}}{...}
}}

\seealso{
Other subscripts: 
\code{\link{mcnebula-class}},
\code{\link{msframe-class}}
}
\concept{subscripts}
