% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-visualize.R
\name{visualize-methods}
\alias{visualize-methods}
\alias{visualize}
\alias{visualize,mcnebula,missing,ANY,missing-method}
\alias{visualize,missing,missing,missing,missing-method}
\alias{visualize,mcnebula,ANY,ANY,ANY-method}
\alias{visualize,mcnebula,character,function,missing-method}
\alias{visualize,mcnebula,numeric,function,missing-method}
\alias{visualize,mcnebula,numeric_or_character,function,logical-method}
\alias{visualize_all,missing,missing,missing,missing-method}
\alias{visualize_all,mcnebula,ANY,ANY,ANY-method}
\alias{visualize_all,mcnebula,logical,function,logical-method}
\title{...}
\usage{
\S4method{visualize}{mcnebula,missing,ANY,missing}(x, fun_modify)

\S4method{visualize}{missing,missing,missing,missing}()

\S4method{visualize}{mcnebula,ANY,ANY,ANY}(x, item, fun_modify, annotate)

\S4method{visualize}{mcnebula,character,`function`,missing}(x, item, fun_modify)

\S4method{visualize}{mcnebula,numeric,`function`,missing}(x, item, fun_modify)

\S4method{visualize}{mcnebula,numeric_or_character,`function`,logical}(x, item, fun_modify, annotate)

\S4method{visualize_all}{missing,missing,missing,missing}()

\S4method{visualize_all}{mcnebula,ANY,ANY,ANY}(x, newpage, fun_modify, legend_hierarchy)

\S4method{visualize_all}{mcnebula,logical,`function`,logical}(x, newpage, fun_modify, legend_hierarchy)
}
\arguments{
\item{x}{...}

\item{fun_modify}{...}

\item{item}{...}

\item{annotate}{...}

\item{newpage}{...}

\item{legend_hierarchy}{...}
}
\description{
...

\code{visualize(x)}: get a 'tbl' about child-nebulae candidates
for \code{visualize} methods to visualize.

\code{visualize()}: get the default parameters for the method
\code{visualize}.

\code{visualize(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{visualize}.

...

\code{visualize_all()}: get the default parameters for the method
\code{visualize_all}.

\code{visualize_all(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{visualize_all}.

...
}
\details{
...

...
}
\examples{
\dontrun{
visualize(...)
}
\dontrun{
visualize_all(...)
}
}
