% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraMethods-workflow.R
\name{workflow-methods}
\alias{workflow-methods}
\alias{workflow}
\alias{workflow,missing,missing,missing,missing,missing,missing-method}
\alias{workflow,ANY,character,ANY,ANY,ANY,ANY-method}
\alias{workflow,numeric,character,environment,character,character,character-method}
\title{Quickly create analysis templates.}
\usage{
\S4method{workflow}{missing,missing,missing,missing,missing,missing}()

\S4method{workflow}{ANY,character,ANY,ANY,ANY,ANY}(sections, mode, envir, sirius_version, sirius_project, ion_mode, ...)

\S4method{workflow}{numeric,character,environment,character,character,character}(sections, mode, envir, sirius_version, sirius_project, ion_mode, ...)
}
\arguments{
\item{sections}{numeric with names. Use \code{workflow()}
to show the available sections.}

\item{mode}{character(1). "print" or "run". If "print", print the
template of the select workflow; If "run", the codes would be eval,
and return with a \linkS4class{report} object.}

\item{envir}{The environment to eval the codes. Default is \code{parent.frame()}.}

\item{sirius_version}{character(1). Passed to \code{\link[=initialize_mcnebula]{initialize_mcnebula()}}.}

\item{sirius_project}{character(1). Passed to \code{\link[=initialize_mcnebula]{initialize_mcnebula()}}.}

\item{ion_mode}{character(1). Set this using \code{ion_mode<-}.}

\item{...}{...}
}
\description{
Quickly create analysis templates and quickly construct a report.
This template contains a number of pre-defined sections, each with a fixed
description and code. The flexible approach is to use this method to obtain the
codes that form these sections and then modify these codes, which is more
likely to result in analysis results and reports that meet the requirements.

\code{workflow()}: get the available section names and
its heading level.

\code{workflow(...)}: use the default parameters whatever 'missing'
while performing the method \code{workflow}.
}
