% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-activate_nebulae.R
\name{activate_nebulae-methods}
\alias{activate_nebulae-methods}
\alias{activate_nebulae}
\alias{activate_nebulae,missing,missing,missing-method}
\alias{activate_nebulae,mcnebula,ANY,ANY-method}
\alias{activate_nebulae,mcnebula,function,function-method}
\alias{ggset_activate_parent_nebula}
\alias{ggset_activate_child_nebulae}
\title{activate Nebulae for visualization}
\usage{
\S4method{activate_nebulae}{missing,missing,missing}()

\S4method{activate_nebulae}{mcnebula,ANY,ANY}(x, fun_default_parent, fun_default_child)

\S4method{activate_nebulae}{mcnebula,`function`,`function`}(x, fun_default_parent, fun_default_child)

ggset_activate_parent_nebula(x)

ggset_activate_child_nebulae(x)
}
\arguments{
\item{x}{\linkS4class{mcnebula} object.}

\item{fun_default_parent}{function. Passed to create \linkS4class{ggset} object
for Parent-Nebula. Default is \code{ggset_activate_parent_nebula}.
Normally not used.}

\item{fun_default_child}{function. Passed to create \linkS4class{ggset} object
for Child-Nebulae. Default is \code{ggset_activate_child_nebulae}.
Normally not used.}
}
\description{
Based on layouts create by \code{\link[=create_parent_layout]{create_parent_layout()}} or \code{\link[=create_child_layouts]{create_child_layouts()}},
use functions to activate Nebulae as \linkS4class{ggset} object for \code{\link[=visualize]{visualize()}}
methods to draw them.

\code{activate_nebulae()}: get the default parameters for the method
\code{activate_nebulae}.

\code{activate_nebulae(x, ...)}:
use the default parameters whatever 'missing'
while performing the method \code{activate_nebulae}.

\code{ggset_activate_parent_nebula}:
create \linkS4class{ggset} object of Parent-Nebula.

\code{ggset_activate_child_nebulae}:
create lists of \linkS4class{ggset} object of Child-Nebulae.
}
\examples{
\dontrun{
activate_nebulae(...)
}
}
\seealso{
\linkS4class{ggset}, \code{\link[=create_parent_layout]{create_parent_layout()}}, \code{\link[=create_child_layouts]{create_child_layouts()}}...
}
