% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraMethods-binary_comparison.R
\name{binary_comparison-methods}
\alias{binary_comparison-methods}
\alias{binary_comparison}
\alias{binary_comparison,missing,missing,missing,missing,missing-method}
\alias{binary_comparison,ANY,ANY,ANY,ANY,ANY-method}
\alias{binary_comparison,ANY,formula,function,ANY,character-method}
\title{Binary comparison for 'features' quantification data}
\usage{
\S4method{binary_comparison}{missing,missing,missing,missing,missing}()

\S4method{binary_comparison}{ANY,ANY,ANY,ANY,ANY}(x, ..., formula, fun_norm, top_coef, contrasts)

\S4method{binary_comparison}{ANY,formula,`function`,ANY,character}(x, formula, fun_norm, top_coef, contrasts)
}
\arguments{
\item{x}{\linkS4class{mcnebula} object.}

\item{...}{expressions, or character strings which can be parsed to
expressions, specifying contrasts. See parameter of \code{...} in
\code{\link[limma:makeContrasts]{limma::makeContrasts()}}.}

\item{formula}{formula. Passed to \code{\link[=model.matrix]{model.matrix()}}.}

\item{fun_norm}{function. For normalization of 'features' quantification
data.}

\item{top_coef}{list, NULL or character(1). Specified the parameter of
\code{coef} in \code{\link[limma:toptable]{limma::topTable()}}. If \code{"all"}, all coefficient in
contrast matrix would be used one by one.}

\item{contrasts}{character vector specifying contrasts.
See parameter \code{contrasts} in \code{\link[limma:makeContrasts]{limma::makeContrasts()}}.}
}
\description{
Use the functions in the 'limma' package for simple binary statistical analysis.

\code{binary_comparison()}:
get the default parameters for the method
\code{binary_comparison}.

\code{binary_comparison(x, ...)}:
use the default parameters whatever 'missing'
while performing the method \code{binary_comparison}.
}
\examples{
\dontrun{
  test <- mcn_5features
  
  ## the previous steps
  test1 <- filter_structure(test)
  test1 <- create_reference(test1)
  test1 <- filter_formula(test1, by_reference = T)
  test1 <- create_features_annotation(test1)
  
  ## set up a simulated quantification data.
  test1 <- .simulate_quant_set(test1)
  ## the simulated data
  features_quantification(test1)
  sample_metadata(test1)
  
  test1 <- binary_comparison(
    test1, control - model,
    model - control, 2 * model - control
  )
  ## see results
  top_table(statistic_set(test1))
  
  ## the default parameters
  binary_comparison()
}
}
\seealso{
\code{\link[stats:model.matrix]{stats::model.matrix()}}, \code{\link[limma:makeContrasts]{limma::makeContrasts()}}, \code{\link[limma:lmFit]{limma::lmFit()}},
\code{\link[limma:ebayes]{limma::eBayes()}}, \code{\link[limma:contrasts.fit]{limma::contrasts.fit()}}, \code{\link[limma:toptable]{limma::topTable()}}...
}
