% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraMethods-collate_data.R
\name{collate_data,ANY,character,function-method}
\alias{collate_data,ANY,character,function-method}
\alias{collate_data}
\alias{collate_data,missing,missing,missing-method}
\alias{collate_data,ANY,ANY,ANY-method}
\alias{read_data,ANY,project_metadata,character,missing,missing,missing,missing,missing-method}
\alias{read_data,missing,missing,character,character,character,character,function,function-method}
\title{...}
\usage{
\S4method{collate_data}{ANY,character,`function`}(x, subscript, fun_collate, ...)

\S4method{collate_data}{missing,missing,missing}()

\S4method{collate_data}{ANY,ANY,ANY}(x, subscript, fun_collate, ...)

\S4method{read_data}{ANY,project_metadata,character,missing,missing,missing,missing,missing}(x, project_metadata, subscript)

\S4method{read_data}{missing,missing,character,character,character,character,`function`,`function`}(subscript, path, .features_id, .candidates_id, fun_read, fun_format)
}
\arguments{
\item{x}{...}

\item{subscript}{...}

\item{fun_collate}{...}

\item{...}{...}

\item{project_metadata}{...}

\item{path}{...}

\item{.features_id}{...}

\item{.candidates_id}{...}

\item{fun_read}{...}

\item{fun_format}{...}
}
\value{
...
}
\description{
...

\code{collate_data()}: get the default parameters for the method
\code{collate_data}.

\code{collate_data(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{collate_data}.

...
}
\details{
...
}
\examples{
\dontrun{
collate_data(...)
}
\dontrun{
read_data(...)
}
}
