% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraMethods-collate_data.R
\name{collate_data-methods}
\alias{collate_data-methods}
\alias{collate_data}
\alias{collate_data,missing,missing,missing-method}
\alias{collate_data,ANY,ANY,ANY-method}
\alias{collate_data,ANY,character,function-method}
\alias{read_data,ANY,project_metadata,character,missing,missing,missing,missing,missing-method}
\alias{read_data,missing,missing,character,character,character,character,function,function-method}
\title{Extract and format data from raw project directory}
\usage{
\S4method{collate_data}{missing,missing,missing}()

\S4method{collate_data}{ANY,ANY,ANY}(x, subscript, fun_collate, ...)

\S4method{collate_data}{ANY,character,`function`}(x, subscript, fun_collate, ...)

\S4method{read_data}{ANY,project_metadata,character,missing,missing,missing,missing,missing}(x, project_metadata, subscript)

\S4method{read_data}{missing,missing,character,character,character,character,`function`,`function`}(subscript, path, .features_id, .candidates_id, fun_read, fun_format)
}
\arguments{
\item{x}{\linkS4class{project} object or other class object inheriting it.}

\item{subscript}{character(1). See \linkS4class{subscript}.}

\item{fun_collate}{function.
Used to extract and format the data from raw project directory.
The default is \code{MCnebula2:::.collate_data.msframe}.}

\item{...}{Other parameters passed to the fun_collate.}

\item{project_metadata}{\linkS4class{project_metadata} object. Specifying the files to read.}

\item{path}{character(1). The path of raw project directory.}

\item{.features_id}{character. ID for signing files in sub-directory of each 'features'.}

\item{.candidates_id}{character. ID for signing each candidates of 'features'.}

\item{fun_read}{function. Used to read files from raw project directory.}

\item{fun_format}{function. Used to format the data.}
}
\description{
The primary method used to extract data from the raw project directory.
By specifying \linkS4class{subscript}, this method reads all corresponding files,
followed by gathering and formating the data, then stores these data in the slot
(\code{dataset(project_dataset(object))}).

\code{collate_data()}: get the default parameters for the method
\code{collate_data}.

\code{collate_data(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{collate_data}.

\code{read_data}: basic methods used to extract and format
data from raw project directory.
}
\details{
This methods requires the name and path of the file in the raw project directory,
as well as the reading function; These are recorded in \linkS4class{project}.
}
\note{
Normally, users do not need to use this method for MCnebula2 analysis.
\code{\link[=filter_formula]{filter_formula()}}, \code{\link[=filter_structure]{filter_structure()}}, \code{\link[=filter_ppcp]{filter_ppcp()}}
provide more understandable usage.
}
\examples{
\dontrun{
collate_data(...)
}
\dontrun{
read_data(...)
}
}
