% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-compute_spectral_similarity.R
\name{compute_spectral_similarity-methods}
\alias{compute_spectral_similarity-methods}
\alias{compute_spectral_similarity}
\alias{compute_spectral_similarity,missing,missing,missing,missing,missing-method}
\alias{compute_spectral_similarity,mcnebula,ANY,ANY,ANY,ANY-method}
\alias{compute_spectral_similarity,missing,missing,missing,lightSpectrum,lightSpectrum-method}
\alias{compute_spectral_similarity,missing,missing,missing,data.frame,data.frame-method}
\alias{compute_spectral_similarity,mcnebula,logical,logical,missing,missing-method}
\title{Compute MS2 spectral similarity}
\usage{
\S4method{compute_spectral_similarity}{missing,missing,missing,missing,missing}()

\S4method{compute_spectral_similarity}{mcnebula,ANY,ANY,ANY,ANY}(x, within_nebula, recompute, sp1, sp2)

\S4method{compute_spectral_similarity}{missing,missing,missing,lightSpectrum,lightSpectrum}(sp1, sp2)

\S4method{compute_spectral_similarity}{missing,missing,missing,data.frame,data.frame}(sp1, sp2)

\S4method{compute_spectral_similarity}{mcnebula,logical,logical,missing,missing}(x, within_nebula, recompute)
}
\arguments{
\item{x}{\linkS4class{mcnebula} object.}

\item{within_nebula}{logical. If \code{TRUE},
only 'features' that exist in a Child-Nebula are compared
for spectral similarity. Data of 'nebula_index' (\code{nebula_index(object)})
would be used for assigning and combining the 'features' of comparison.}

\item{recompute}{logical. If \code{TRUE}, discard the existing data in the object,
and recompute the spectral similarity.}

\item{sp1}{data.frame. An additional channel for comparing two spectrum.
Contains 'mz' and 'intensity' for spectral comparison.}

\item{sp2}{data.frame. An additional channel for comparing two spectrum.
Contains 'mz' and 'intensity' for spectral comparison.}
}
\description{
These are methods stripped from \link[MSnbase:compareSpectra-methods]{MSnbase::compareSpectra}.
The unnecessary parts were removed, \code{\link[=compute_spectral_similarity]{compute_spectral_similarity()}} only
calculate the 'dotproduct' for two spectra and get the similarity value.
It allows faster results for batch comparisons.

\code{compute_spectral_similarity()}: get the default parameters for the method
\code{compute_spectral_similarity}.

\code{compute_spectral_similarity(x, ...)}:
use the default parameters whatever 'missing'
while performing the method \code{compute_spectral_similarity}.
}
\examples{
\dontrun{
compute_spectral_similarity(...)
}
}
\seealso{
\code{\link[MSnbase:compareSpectra-methods]{MSnbase::compareSpectra()}}.
}
