% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-compute_spectral_similarity.R
\name{compute_spectral_similarity,mcnebula,logical,logical,missing,missing-method}
\alias{compute_spectral_similarity,mcnebula,logical,logical,missing,missing-method}
\alias{compute_spectral_similarity}
\alias{compute_spectral_similarity,missing,missing,missing,missing,missing-method}
\alias{compute_spectral_similarity,mcnebula,ANY,ANY,ANY,ANY-method}
\alias{compute_spectral_similarity,missing,missing,missing,lightSpectrum,lightSpectrum-method}
\alias{compute_spectral_similarity,missing,missing,missing,data.frame,data.frame-method}
\title{...}
\usage{
\S4method{compute_spectral_similarity}{mcnebula,logical,logical,missing,missing}(x, within_nebula, recompute)

\S4method{compute_spectral_similarity}{missing,missing,missing,missing,missing}()

\S4method{compute_spectral_similarity}{mcnebula,ANY,ANY,ANY,ANY}(x, within_nebula, recompute, sp1, sp2)

\S4method{compute_spectral_similarity}{missing,missing,missing,lightSpectrum,lightSpectrum}(sp1, sp2)

\S4method{compute_spectral_similarity}{missing,missing,missing,data.frame,data.frame}(sp1, sp2)
}
\arguments{
\item{x}{...}

\item{within_nebula}{...}

\item{recompute}{...}

\item{sp1}{...}

\item{sp2}{...}
}
\value{
...
}
\description{
...

\code{compute_spectral_similarity()}: get the default parameters for the method
\code{compute_spectral_similarity}.

\code{compute_spectral_similarity(x, ...)}:
use the default parameters whatever 'missing'
while performing the method \code{compute_spectral_similarity}.
}
\details{
...
}
\examples{
\dontrun{
compute_spectral_similarity(...)
}
}
\seealso{
\code{\link[=fun]{fun()}}
}
