% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-create_child_layouts.R
\name{create_child_layouts-methods}
\alias{create_child_layouts-methods}
\alias{create_child_layouts}
\alias{create_child_layouts,missing,missing,missing,missing,missing,missing,missing-method}
\alias{create_child_layouts,mcnebula,ANY,ANY,ANY,ANY,ANY,ANY-method}
\title{...}
\usage{
\S4method{create_child_layouts}{missing,missing,missing,missing,missing,missing,missing}()

\S4method{create_child_layouts}{mcnebula,ANY,ANY,ANY,ANY,ANY,ANY}(
  x,
  ggraph_layouts,
  seeds,
  grid_layout,
  viewports,
  panel_viewport,
  legend_viewport
)
}
\arguments{
\item{x}{...}

\item{ggraph_layouts}{...}

\item{seeds}{...}

\item{grid_layout}{...}

\item{viewports}{...}

\item{panel_viewport}{...}

\item{legend_viewport}{...}
}
\value{
...
}
\description{
...

\code{create_child_layouts()}: get the function for generating
default parameters for the method
\code{create_child_layouts}.

...

\code{create_child_layouts(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{create_child_layouts}.
}
\details{
...

...
}
\examples{
\dontrun{
create_child_layouts(...)
}
}
