% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-create_child_nebulae.R
\name{create_child_nebulae,mcnebula,numeric,numeric,logical-method}
\alias{create_child_nebulae,mcnebula,numeric,numeric,logical-method}
\alias{create_child_nebulae}
\alias{create_child_nebulae,missing,missing,missing,missing-method}
\alias{create_child_nebulae,mcnebula,ANY,ANY,ANY-method}
\title{...}
\usage{
\S4method{create_child_nebulae}{mcnebula,numeric,numeric,logical}(x, edge_cutoff, max_edge_number, use_tracer)

\S4method{create_child_nebulae}{missing,missing,missing,missing}()

\S4method{create_child_nebulae}{mcnebula,ANY,ANY,ANY}(x, edge_cutoff, max_edge_number, use_tracer)
}
\arguments{
\item{x}{...}

\item{edge_cutoff}{...}

\item{max_edge_number}{...}
}
\value{
...
}
\description{
...

\code{create_child_nebulae()}: get the default parameters for the method
\code{create_child_nebulae}.

\code{create_child_nebulae(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{create_child_nebulae}.
}
\details{
...
}
\examples{
\dontrun{
create_child_nebulae(...)
}
}
