% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-create_features_annotation.R
\name{create_features_annotation-methods}
\alias{create_features_annotation-methods}
\alias{create_features_annotation}
\alias{create_features_annotation,mcnebula,missing,missing-method}
\alias{create_features_annotation,mcnebula,data.frame,numeric-method}
\alias{create_features_annotation,mcnebula,data.frame,missing-method}
\title{...}
\usage{
\S4method{create_features_annotation}{mcnebula,missing,missing}(x)

\S4method{create_features_annotation}{mcnebula,data.frame,numeric}(x, extra_data, column)

\S4method{create_features_annotation}{mcnebula,data.frame,missing}(x, extra_data)
}
\arguments{
\item{x}{...}

\item{extra_data}{...}

\item{column}{...}
}
\value{
...
}
\description{
...

...
}
\details{
...

...
}
\examples{
\dontrun{
create_features_annotation(...)
}
}
