% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-create_hierarchy.R
\name{create_hierarchy-methods}
\alias{create_hierarchy-methods}
\alias{create_hierarchy}
\alias{create_hierarchy,missing,missing-method}
\alias{create_hierarchy,mcnebula,ANY-method}
\alias{create_hierarchy,mcnebula,function-method}
\title{Create hierarchy data of chemical classification}
\usage{
\S4method{create_hierarchy}{missing,missing}()

\S4method{create_hierarchy}{mcnebula,ANY}(x, fun_organize)

\S4method{create_hierarchy}{mcnebula,`function`}(x, fun_organize)
}
\arguments{
\item{x}{\linkS4class{mcnebula} object.}

\item{fun_organize}{function. Normally not used.
Default is \code{MCnebula2:::.build_hierarchy}.}
}
\description{
Methods used to create hierarchy data of chemical classification.
Annotate all chemical classes with hierarchy number.

\code{create_hierarchy()}:
get the default parameters for the method
\code{create_hierarchy}.

\code{create_hierarchy(x, ...)}:
use the default parameters whatever 'missing'
while performing the method \code{create_hierarchy}.
}
\examples{
\dontrun{
create_hierarchy(...)
}
}
