% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-create_nebula_index.R
\name{create_nebula_index,mcnebula,logical-method}
\alias{create_nebula_index,mcnebula,logical-method}
\alias{create_nebula_index}
\alias{create_nebula_index,missing,missing-method}
\alias{create_nebula_index,mcnebula,ANY-method}
\title{...}
\usage{
\S4method{create_nebula_index}{mcnebula,logical}(x, force)

\S4method{create_nebula_index}{missing,missing}()

\S4method{create_nebula_index}{mcnebula,ANY}(x, force)
}
\arguments{
\item{x}{...}

\item{force}{...}
}
\value{
...
}
\description{
...

\code{create_nebula_index()}: get the default parameters for the method
\code{create_nebula_index}.

\code{create_nebula_index(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{create_nebula_index}.
}
\details{
...
}
\examples{
\dontrun{
create_nebula_index(...)
}
}
