% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-create_parent_nebula.R
\name{create_parent_nebula,mcnebula,numeric,logical-method}
\alias{create_parent_nebula,mcnebula,numeric,logical-method}
\alias{create_parent_nebula}
\alias{create_parent_nebula,mcnebula,missing,missing-method}
\alias{create_parent_nebula,mcnebula,numeric,missing-method}
\title{...}
\usage{
\S4method{create_parent_nebula}{mcnebula,numeric,logical}(x, edge_cutoff, remove_isolate)

\S4method{create_parent_nebula}{mcnebula,missing,missing}(x)

\S4method{create_parent_nebula}{mcnebula,numeric,missing}(x, edge_cutoff)
}
\arguments{
\item{x}{...}

\item{edge_cutoff}{...}

\item{remove_isolate}{...}
}
\value{
...
}
\description{
...

\code{create_parent_nebula()}: get the default parameters for the method
\code{create_parent_nebula}.

\code{create_parent_nebula(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{create_parent_nebula}.
}
\details{
...
}
\examples{
\dontrun{
create_parent_nebula(...)
}
}
\seealso{
\code{\link[=fun]{fun()}}
}
