% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-create_reference.R
\name{create_reference-methods}
\alias{create_reference-methods}
\alias{create_reference}
\alias{create_reference,mcnebula,ANY,ANY,ANY,ANY,logical,ANY-method}
\alias{create_reference,mcnebula,missing,missing,missing,missing,missing,missing-method}
\alias{create_reference,mcnebula,character,missing,missing,missing,missing,missing-method}
\alias{create_reference,mcnebula,missing,character,missing,missing,missing,missing-method}
\alias{create_reference,mcnebula,missing,missing,data.frame,character,missing,missing-method}
\alias{create_reference,mcnebula,missing,missing,data.frame,integer,missing,missing-method}
\alias{create_reference,mcnebula,missing,missing,data.frame,missing,missing,missing-method}
\title{...}
\usage{
\S4method{create_reference}{mcnebula,ANY,ANY,ANY,ANY,logical,ANY}(x, from, subscript, data, columns, fill, MoreArgs)

\S4method{create_reference}{mcnebula,missing,missing,missing,missing,missing,missing}(x)

\S4method{create_reference}{mcnebula,character,missing,missing,missing,missing,missing}(x, from)

\S4method{create_reference}{mcnebula,missing,character,missing,missing,missing,missing}(x, subscript)

\S4method{create_reference}{mcnebula,missing,missing,data.frame,character,missing,missing}(x, data, columns)

\S4method{create_reference}{mcnebula,missing,missing,data.frame,integer,missing,missing}(x, data, columns)

\S4method{create_reference}{mcnebula,missing,missing,data.frame,missing,missing,missing}(x, data)
}
\arguments{
\item{x}{...}

\item{from}{...}

\item{subscript}{...}

\item{data}{...}

\item{columns}{...}

\item{fill}{...}

\item{MoreArgs}{...}
}
\value{
...
}
\description{
...

...
}
\details{
...

...
}
\examples{
\dontrun{
create_reference(...)
}
}
