% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cross_filter_stardust.R
\name{cross_filter_stardust-methods}
\alias{cross_filter_stardust-methods}
\alias{cross_filter_stardust}
\alias{cross_filter_stardust,missing-method}
\alias{cross_filter_stardust,mcnebula-method}
\alias{cross_filter_quantity,mcnebula,numeric,numeric-method}
\alias{cross_filter_quantity}
\alias{cross_filter_score,mcnebula,character,numeric,numeric-method}
\alias{cross_filter_score}
\alias{cross_filter_identical,mcnebula,numeric,numeric-method}
\alias{cross_filter_identical}
\title{...}
\usage{
\S4method{cross_filter_stardust}{missing}()

\S4method{cross_filter_stardust}{mcnebula}(
  x,
  min_number,
  max_ratio,
  types,
  cutoff,
  tolerance,
  hierarchy_range,
  identical_factor
)

\S4method{cross_filter_quantity}{mcnebula,numeric,numeric}(x, min_number, max_ratio)

\S4method{cross_filter_score}{mcnebula,character,numeric,numeric}(x, types, cutoff, tolerance)

\S4method{cross_filter_identical}{mcnebula,numeric,numeric}(x, hierarchy_range, identical_factor)
}
\arguments{
\item{x}{...}

\item{min_number}{...}

\item{max_ratio}{...}

\item{types}{...}

\item{cutoff}{...}

\item{tolerance}{...}

\item{hierarchy_range}{...}

\item{identical_factor}{...

\code{cross_filter_score}, \code{cross_filter_identical}.}
}
\value{
...

...

...
}
\description{
...

\code{cross_filter_stardust()}: get the default parameters for the method
\code{cross_filter_stardust}.

\code{cross_filter_stardust(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{cross_filter_stardust}.

\code{cross_filter_stardust} include 3 parts: \code{cross_filter_quantity},

...

...

...
}
\details{
...

...

...

...
}
\examples{
\dontrun{
cross_filter_quantity(...)
}
\dontrun{
cross_filter_score(...)
}
\dontrun{
cross_filter_identical(...)
}
}
