% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraMethods-draw_structures.R
\name{draw_structures,mcnebula,character-method}
\alias{draw_structures,mcnebula,character-method}
\alias{draw_structures}
\alias{show_structure,ANY,character-method}
\title{...}
\usage{
\S4method{draw_structures}{mcnebula,character}(x, nebula_name)

\S4method{show_structure}{ANY,character}(x, .features_id)
}
\arguments{
\item{x}{...}

\item{nebula_name}{...}

\item{.features_id}{...}
}
\value{
...
}
\description{
...

...
}
\details{
...
}
\examples{
\dontrun{
draw_structures(...)
}
\dontrun{
show_structure(...)
}
}
