% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraMethods-draw_structures.R
\name{draw_structures-methods}
\alias{draw_structures-methods}
\alias{draw_structures}
\alias{draw_structures,mcnebula,character-method}
\alias{show_structure,ANY,character-method}
\title{Draw and visualize chemcial structure}
\usage{
\S4method{draw_structures}{mcnebula,character}(x, nebula_name)

\S4method{show_structure}{ANY,character}(x, .features_id)
}
\arguments{
\item{x}{\linkS4class{mcnebula} object.}

\item{nebula_name}{character(1). Chemical classes in 'nebula_index' data.
Specified to draw chemical structures of all the 'features' of that.}

\item{.features_id}{character(1). The ID of 'features' to show the
chemical structure.}
}
\description{
Methods used for drawing and visualizing chemical structures of 'features'
in Child-Nebulae.
\code{\link[ChemmineOB:convertToImage]{ChemmineOB::convertToImage()}} is the core function used for drawing chemical
structures.

\code{show_structure}: visualize the chemical structure of 'feature'
which has been drawn.
}
\examples{
\dontrun{
  test <- mcn_5features
  
  ## the previous steps
  test1 <- filter_structure(test)
  test1 <- create_reference(test1)
  test1 <- filter_formula(test1, by_reference = T)
  test1 <- create_stardust_classes(test1)
  test1 <- create_features_annotation(test1)
  test1 <- cross_filter_stardust(test1, 2, 1)
  test1 <- create_nebula_index(test1)
  test1 <- compute_spectral_similarity(test1)
  test1 <- create_child_nebulae(test1, 0.01, 5)
  test1 <- create_child_layouts(test1)
  test1 <- activate_nebulae(test1)
  
  ## optional 'nebula_name'
  visualize(test1)
  ## a class for example
  class <- visualize(test1)$class.name[1]
  tmp <- export_path(test1)
  test1 <- draw_structures(test1, class)
  
  ## see results
  grobs <- structures_grob(child_nebulae(test1))
  grobs
  grid::grid.draw(grobs[[1]])
  ## visualize with ID of 'feature' (.features_id)
  ids <- names(grobs)
  show_structure(test1, ids[1])
  
  unlink(tmp, T, T)
}
}
\seealso{
\code{\link[ChemmineOB:convertToImage]{ChemmineOB::convertToImage()}}.
}
