% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-filter_ppcp.R
\name{filter_ppcp-methods}
\alias{filter_ppcp-methods}
\alias{filter_ppcp}
\alias{filter_ppcp,missing,missing,missing-method}
\alias{filter_ppcp,mcnebula,ANY,ANY-method}
\alias{filter_ppcp,mcnebula,function,logical-method}
\title{...}
\usage{
\S4method{filter_ppcp}{missing,missing,missing}()

\S4method{filter_ppcp}{mcnebula,ANY,ANY}(x, fun_filter, ..., by_reference)

\S4method{filter_ppcp}{mcnebula,`function`,logical}(x, fun_filter, ..., by_reference)
}
\arguments{
\item{param}{...}
}
\value{
...
}
\description{
...

\code{filter_ppcp()}: get the default parameters for the method
\code{filter_ppcp}.

\code{filter_ppcp(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{filter_ppcp}.

...
}
\details{
...

...
}
\examples{
\dontrun{
filter_ppcp(...)
}
}
