% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-filter_structure.R
\name{filter_structure,missing,missing,missing-method}
\alias{filter_structure,missing,missing,missing-method}
\alias{filter_structure,mcnebula,ANY,ANY-method}
\alias{filter_structure,mcnebula,function,logical-method}
\alias{filter_structure}
\title{...}
\usage{
\S4method{filter_structure}{missing,missing,missing}()

\S4method{filter_structure}{mcnebula,ANY,ANY}(x, fun_filter, ..., by_reference)

\S4method{filter_structure}{mcnebula,`function`,logical}(x, fun_filter, ..., by_reference)
}
\arguments{
\item{x}{\linkS4class{mcnebula} object.}

\item{fun_filter}{function. e.g., \code{\link[dplyr:filter]{dplyr::filter()}}, \code{\link[dplyr:slice]{dplyr::slice()}} \code{\link[=head]{head()}}...}

\item{...}{parameters pass to 'fun_filter'.}

\item{by_reference}{...}
}
\value{
...
}
\description{
\code{filter_structure()}: get the default parameters for the method
\code{filter_structure}.

\code{filter_structure(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{filter_structure}.

...
}
\details{
...
}
\examples{
\dontrun{
filter_structure(...)
}
}
