% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-modify_ggset.R
\name{fun_modify}
\alias{fun_modify}
\alias{modify_default_child}
\alias{modify_set_labs_and_unify_scale_limits}
\alias{modify_annotate_child}
\alias{modify_rm_legend}
\alias{modify_set_margin}
\alias{modify_unify_scale_limits}
\alias{modify_set_labs}
\title{...}
\usage{
modify_default_child(ggset, x)

modify_set_labs_and_unify_scale_limits(ggset, x)

modify_annotate_child(ggset, x)

modify_rm_legend(ggset)

modify_set_margin(ggset, margin = grid::unit(rep(-8, 4), "lines"))

modify_unify_scale_limits(ggset, x)

modify_set_labs(ggset, x)
}
\arguments{
\item{ggset}{...}

\item{x}{...}
}
\description{
...

\code{modify_default_child}: ...

\code{modify_set_labs_and_unify_scale_limits}: ...

\code{modify_annotate_child}: ...

\code{modify_rm_legend}: ...

\code{modify_set_margin}: ...

\code{modify_unify_scale_limits}: ...

\code{modify_set_labs}: ...
}
\details{
...
}
