% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ggset.R
\docType{class}
\name{ggset-class}
\alias{ggset-class}
\alias{.ggset}
\alias{ggset}
\alias{show_layers,ggset-method}
\alias{show_layers}
\alias{new_ggset,ANY-method}
\alias{new_ggset}
\alias{mutate_layer,ggset,numeric-method}
\alias{mutate_layer}
\alias{mutate_layer,ggset,character-method}
\alias{call_command,ggset-method}
\alias{call_command}
\title{...}
\usage{
\S4method{show_layers}{ggset}(x)

\S4method{new_ggset}{ANY}(...)

\S4method{mutate_layer}{ggset,numeric}(x, layer, ...)

\S4method{mutate_layer}{ggset,character}(x, layer, ...)

\S4method{call_command}{ggset}(x)
}
\arguments{
\item{x}{...}

\item{...}{...}

\item{layer}{...}
}
\description{
...

\code{show_layers}: ...

\code{new_ggset}: ...

\code{mutate_layer}: ...

\code{call_command}: plot as 'ggplot' object.
}
\section{Slots}{

\describe{
\item{\code{layers}}{...}
}}

\examples{
\dontrun{
new('ggset', ...)
}
\dontrun{
show_layers(...)
}
\dontrun{
new_ggset(...)
}
\dontrun{
mutate_layer(...)
}
}
\seealso{
Other layerSets: 
\code{\link{layerSet-class}},
\code{\link{report-class}}

Other call_commands: 
\code{\link{code_block-class}},
\code{\link{report-class}},
\code{\link{section-class}}
}
\concept{call_commands}
\concept{layerSets}
