% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ggset.R
\docType{class}
\name{ggset-class}
\alias{ggset-class}
\alias{.ggset}
\alias{ggset}
\alias{show_layers,ggset-method}
\alias{show_layers}
\alias{new_ggset,ANY-method}
\alias{new_ggset}
\alias{mutate_layer,ggset,numeric-method}
\alias{mutate_layer}
\alias{mutate_layer,ggset,character-method}
\alias{call_command,ggset-method}
\alias{call_command}
\title{Management for 'ggplot' visualzation}
\usage{
\S4method{show_layers}{ggset}(x)

\S4method{new_ggset}{ANY}(...)

\S4method{mutate_layer}{ggset,numeric}(x, layer, ...)

\S4method{mutate_layer}{ggset,character}(x, layer, ...)

\S4method{call_command}{ggset}(x)
}
\arguments{
\item{x}{\linkS4class{ggset} object}

\item{...}{parameters passed to the layer.}

\item{layer}{numeric(1) or character(1). If "character", the name must be unique
in slot \code{layers}.}
}
\description{
Let each packed "ggplot2" function (packed as \linkS4class{command} object)
into layers in sequence, allowing post modifications programmatically
and visualizing as "ggplot2" plot at any time.

\code{show_layers}: show functions and parameters in layers
with a pretty and readable form.

\code{new_ggset}: Simplified creation of \linkS4class{ggset} object.

\code{mutate_layer}:
Pass new parameters or modify pre-existing parameters to the packed function.

\code{call_command}: plot as 'ggplot' object.
}
\section{Slots}{

\describe{
\item{\code{layers}}{list with names. Each element of list must be a \linkS4class{command} object
packed 'ggplot2' function and its args.}
}}

\examples{
\dontrun{
  data <- data.frame(x = 1:10, y = 1:10)
  layer1 <- new_command(ggplot, data)
  layer2 <- new_command(geom_point, aes(x = x, y = y))
  layer3 <- new_command(labs, x = "x label", y = "y label")
  layer4 <- new_command(theme, text = element_text(family = "Times"))
  
  ## gather
  ggset <- new_ggset(layer1, layer2, layer3, layer4)
  ggset
  ## visualize
  p <- call_command(ggset)
  p
  
  ## add layers
  layer5 <- new_command(
    geom_text,
    aes(x = x, y = y, label = paste0("label_", x))
  )
  layer6 <- new_command(ggtitle, "this is title")
  ggset <- add_layers(ggset, layer5, layer6)
  call_command(ggset)
  
  ## delete layers
  ggset <- delete_layers(ggset, 5:6)
  call_command(ggset)
  
  ## mutate layer
  ggset <- mutate_layer(ggset, "theme",
    legend.position = "none",
    plot.background = element_rect(fill = "red")
  )
  ggset <- mutate_layer(ggset, "geom_point",
    mapping = aes(x = x, y = y, color = x)
  )
  call_command(ggset)
}
}
\seealso{
Other layerSets: 
\code{\link{layerSet-class}},
\code{\link{report-class}}

Other call_commands: 
\code{\link{code_block-class}},
\code{\link{command-class}},
\code{\link{report-class}},
\code{\link{section-class}}
}
\concept{call_commands}
\concept{layerSets}
