% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraMethods-report.R
\name{history_rblock,numeric,missing,missing,numeric-method}
\alias{history_rblock,numeric,missing,missing,numeric-method}
\alias{history_rblock}
\alias{history_rblock,missing,missing,missing,missing-method}
\alias{history_rblock,numeric,ANY,ANY,ANY-method}
\alias{history_rblock,missing,character,character,ANY-method}
\title{...}
\usage{
\S4method{history_rblock}{numeric,missing,missing,numeric}(nrow, exclude)

\S4method{history_rblock}{missing,missing,missing,missing}()

\S4method{history_rblock}{numeric,ANY,ANY,ANY}(nrow, pattern_start, pattern_end, exclude)

\S4method{history_rblock}{missing,character,character,ANY}(pattern_start, pattern_end, exclude)
}
\arguments{
\item{nrow}{...}

\item{exclude}{...}

\item{pattern_start}{...}

\item{pattern_end}{...}
}
\value{
...
}
\description{
...

\code{history_rblock()}: get the default parameters for the method
\code{history_rblock}.

\code{history_rblock(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{history_rblock}.
}
\details{
...
}
\examples{
\dontrun{
history_rblock(...)
}
}
\seealso{
\code{\link{report-class}}, \code{\link{code_block-class}}...
}
