% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraMethods-report.R
\name{include_figure-methods}
\alias{include_figure-methods}
\alias{include_figure}
\alias{include_figure,character,character,character-method}
\title{...}
\usage{
\S4method{include_figure}{character,character,character}(file, name, caption)
}
\arguments{
\item{file}{...}

\item{name}{...}

\item{caption}{...}
}
\value{
...
}
\description{
...

...
}
\details{
...

...
}
\examples{
\dontrun{
include_figure(...)
}
}
\seealso{
\code{\link{report-class}}, \code{\link{code_block-class}}...
}
