% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraMethods-report.R
\name{include_table,data.frame,character,character-method}
\alias{include_table,data.frame,character,character-method}
\alias{include_table}
\title{...}
\usage{
\S4method{include_table}{data.frame,character,character}(data, name, caption)
}
\arguments{
\item{data}{...}

\item{name}{...}

\item{caption}{...}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
\dontrun{
include_table(...)
}
}
\seealso{
\code{\link[=fun]{fun()}}
}
