% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-initialize_mcnebula.R
\name{initialize_mcnebula,mcnebula,ANY-method}
\alias{initialize_mcnebula,mcnebula,ANY-method}
\alias{initialize_mcnebula}
\alias{initialize_mcnebula,melody,ANY-method}
\title{...}
\usage{
\S4method{initialize_mcnebula}{mcnebula,ANY}(x, sirius_version, sirius_project, output_directory)

\S4method{initialize_mcnebula}{melody,ANY}(x)
}
\arguments{
\item{x}{\linkS4class{melody} object.}

\item{sirius_version}{character. e.g., "sirius.v4".}

\item{sirius_project}{character. The path of SIRIUS project space.}

\item{output_directory}{character. The path for output.}
}
\value{
\linkS4class{mcnebula} object.

\linkS4class{melody} object.
}
\description{
...

...
}
\details{
...

...
}
\examples{
\dontrun{
initialize_mcnebula(...)
}
\dontrun{
initialize_mcnebula(...)
}
}
\seealso{
\code{\link[ggsci:pal_simpsons]{ggsci::pal_simpsons()}}, \code{\link[ggsci:pal_igv]{ggsci::pal_igv()}}, \code{\link[ggsci:pal_ucscgb]{ggsci::pal_ucscgb()}},
\code{\link[ggsci:pal_d3]{ggsci::pal_d3()}}...
}
