% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mcn_dataset.R
\docType{class}
\name{mcn_dataset-class}
\alias{mcn_dataset-class}
\alias{.mcn_dataset}
\alias{mcn_dataset}
\alias{mcn_dataset,ANY-method}
\alias{mcn_dataset<-,ANY-method}
\alias{mcn_dataset<-}
\alias{latest,mcn_dataset,ANY,ANY-method}
\alias{latest}
\alias{extract_mcnset,ANY,character-method}
\alias{extract_mcnset}
\title{Store processed data}
\usage{
\S4method{mcn_dataset}{ANY}(x)

\S4method{mcn_dataset}{ANY}(x) <- value

\S4method{latest}{mcn_dataset,ANY,ANY}(x)

\S4method{extract_mcnset}{ANY,character}(x, subscript)
}
\arguments{
\item{value}{The value for the slot.}

\item{subscript}{See \linkS4class{subscript}}
}
\description{
This is a class object used to store filtered data and formated data.
These data would be used for further analysis or visualization.

\code{mcn_dataset}, \code{mcn_dataset<-}: getter and setter
for the \code{mcn_dataset} slot of the object.

\code{latest}: get the first data in \code{dataset} slot and
format as "tbl". Equals:
\itemize{
\item \code{latest(object)}
\item \code{tibble::as_tibble(entity(dataset(x)[[1]]))}.
}

\code{extract_mcnset}: For fast extract data in object which containing
\code{mcn_dataset} slot. Normally not used.
}
\section{Slots}{

\describe{
\item{\code{dataset}}{list with names of \linkS4class{subscript}. Store preliminary filtered data.}

\item{\code{reference}}{list with names of standard names. Store formated data, which is useful
reference for further analysis or visualization.}

\item{\code{backtrack}}{list with names. Recovery stations halfway through data processing.}
}}

\examples{
\dontrun{
new('mcn_dataset', ...)
}
\dontrun{
latest(...)
}
\dontrun{
extract_mcnset(...)
}
}
\seealso{
\linkS4class{dataset}

\linkS4class{subscript}

Other datasets: 
\code{\link{dataset-class}},
\code{\link{project_dataset-class}}

Other latests: 
\code{\link{mcnebula-class}},
\code{\link{msframe-class}},
\code{\link{project_dataset-class}},
\code{\link{project_metadata-class}}
}
\concept{datasets}
\concept{latests}
