% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-msframe.R
\docType{class}
\name{msframe-class}
\alias{msframe-class}
\alias{.msframe}
\alias{msframe}
\alias{show,msframe-method}
\alias{show}
\alias{msframe,ANY-method}
\alias{msframe<-,ANY-method}
\alias{msframe<-}
\alias{latest,msframe,ANY,ANY-method}
\alias{latest}
\alias{entity,msframe-method}
\alias{entity}
\alias{entity<-,msframe-method}
\alias{entity<-}
\alias{format_msframe,msframe,missing,missing,missing,missing,function-method}
\alias{format_msframe}
\alias{format_msframe,data.frame,missing,missing,missing,missing,function-method}
\alias{format_msframe,msframe,character,missing,character,missing,missing-method}
\alias{format_msframe,msframe,missing,missing,missing,missing,missing-method}
\alias{format_msframe,msframe,missing,function,missing,function,missing-method}
\alias{filter_msframe,msframe,function,missing-method}
\alias{filter_msframe}
\alias{filter_msframe,msframe,function,formula-method}
\title{...}
\usage{
\S4method{show}{msframe}(object)

\S4method{msframe}{ANY}(x)

\S4method{msframe}{ANY}(x) <- value

\S4method{latest}{msframe,ANY,ANY}(x)

\S4method{entity}{msframe}(x)

\S4method{entity}{msframe}(x) <- value

\S4method{format_msframe}{msframe,missing,missing,missing,missing,`function`}(x, fun_format)

\S4method{format_msframe}{data.frame,missing,missing,missing,missing,`function`}(x, fun_format)

\S4method{format_msframe}{msframe,character,missing,character,missing,missing}(x, names, types)

\S4method{format_msframe}{msframe,missing,missing,missing,missing,missing}(x)

\S4method{format_msframe}{msframe,missing,`function`,missing,`function`,missing}(x, fun_names, fun_types)

\S4method{filter_msframe}{msframe,`function`,missing}(x, fun_filter, f, ...)

\S4method{filter_msframe}{msframe,`function`,formula}(x, fun_filter, f, ...)
}
\arguments{
\item{x}{...}

\item{value}{The value for the slot.}

\item{fun_format}{...}

\item{names}{...}

\item{types}{...}

\item{fun_names}{...}

\item{fun_types}{...}

\item{fun_filter}{...}

\item{f}{...}

\item{...}{...}
}
\value{
...
}
\description{
...

\code{msframe}, \code{msframe<-}: getter and setter
for the \code{msframe} slot of the object.

\code{latest}: get data inside \code{entity(object)} and format as
'tbl'.

\code{entity}, \code{entity<-}: getter and setter
for the \code{entity} slot of the object.

...

...
}
\details{
...

...
}
\section{Slots}{

\describe{
\item{\code{entity}}{...}

\item{\code{subscript}}{...}
}}

\examples{
\dontrun{
new('msframe', ...)
}
\dontrun{
latest(...)
}
\dontrun{
format_msframe(...)
}
\dontrun{
filter_msframe(...)
}
}
\seealso{
\code{\link[tibble:as_tibble]{tibble::as_tibble()}}

Other subscripts: 
\code{\link{mcnebula-class}},
\code{\link{subscript-class}}

Other latests: 
\code{\link{mcn_dataset-class}},
\code{\link{mcnebula-class}},
\code{\link{project_dataset-class}}
}
\concept{latests}
\concept{subscripts}
