% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-nebula.R
\docType{class}
\name{nebula-class}
\alias{nebula-class}
\alias{.nebula}
\alias{nebula}
\alias{parent_nebula-class}
\alias{.parent_nebula}
\alias{parent_nebula}
\alias{child_nebulae-class}
\alias{.child_nebulae}
\alias{child_nebulae}
\alias{show,parent_nebula-method}
\alias{show,child_nebulae-method}
\alias{parent_nebula,ANY-method}
\alias{parent_nebula<-,ANY-method}
\alias{parent_nebula<-}
\alias{child_nebulae,ANY-method}
\alias{child_nebulae<-,ANY-method}
\alias{child_nebulae<-}
\alias{igraph,ANY-method}
\alias{igraph}
\alias{igraph<-,ANY-method}
\alias{igraph<-}
\alias{tbl_graph,ANY-method}
\alias{tbl_graph}
\alias{tbl_graph<-,ANY-method}
\alias{tbl_graph<-}
\alias{layout_ggraph,ANY-method}
\alias{layout_ggraph}
\alias{layout_ggraph<-,ANY-method}
\alias{layout_ggraph<-}
\alias{grid_layout,ANY-method}
\alias{grid_layout}
\alias{grid_layout<-,ANY-method}
\alias{grid_layout<-}
\alias{viewports,ANY-method}
\alias{viewports}
\alias{viewports<-,ANY-method}
\alias{viewports<-}
\alias{ggset,ANY-method}
\alias{ggset}
\alias{ggset<-,ANY-method}
\alias{ggset<-}
\alias{panel_viewport,ANY-method}
\alias{panel_viewport}
\alias{panel_viewport<-,ANY-method}
\alias{panel_viewport<-}
\alias{legend_viewport,ANY-method}
\alias{legend_viewport}
\alias{legend_viewport<-,ANY-method}
\alias{legend_viewport<-}
\alias{structures_grob,ANY-method}
\alias{structures_grob}
\alias{structures_grob<-,ANY-method}
\alias{structures_grob<-}
\alias{nodes_ggset,ANY-method}
\alias{nodes_ggset}
\alias{nodes_ggset<-,ANY-method}
\alias{nodes_ggset<-}
\alias{nodes_grob,ANY-method}
\alias{nodes_grob}
\alias{nodes_grob<-,ANY-method}
\alias{nodes_grob<-}
\alias{ppcp_data,ANY-method}
\alias{ppcp_data}
\alias{ppcp_data<-,ANY-method}
\alias{ppcp_data<-}
\alias{ration_data,ANY-method}
\alias{ration_data}
\alias{ration_data<-,ANY-method}
\alias{ration_data<-}
\alias{ggset_annotate,ANY-method}
\alias{ggset_annotate}
\alias{ggset_annotate<-,ANY-method}
\alias{ggset_annotate<-}
\title{Visualization component of chemical Nebulae/Nebula}
\usage{
\S4method{show}{parent_nebula}(object)

\S4method{show}{child_nebulae}(object)

\S4method{parent_nebula}{ANY}(x)

\S4method{parent_nebula}{ANY}(x) <- value

\S4method{child_nebulae}{ANY}(x)

\S4method{child_nebulae}{ANY}(x) <- value

\S4method{igraph}{ANY}(x)

\S4method{igraph}{ANY}(x) <- value

\S4method{tbl_graph}{ANY}(x)

\S4method{tbl_graph}{ANY}(x) <- value

\S4method{layout_ggraph}{ANY}(x)

\S4method{layout_ggraph}{ANY}(x) <- value

\S4method{grid_layout}{ANY}(x)

\S4method{grid_layout}{ANY}(x) <- value

\S4method{viewports}{ANY}(x)

\S4method{viewports}{ANY}(x) <- value

\S4method{ggset}{ANY}(x)

\S4method{ggset}{ANY}(x) <- value

\S4method{panel_viewport}{ANY}(x)

\S4method{panel_viewport}{ANY}(x) <- value

\S4method{legend_viewport}{ANY}(x)

\S4method{legend_viewport}{ANY}(x) <- value

\S4method{structures_grob}{ANY}(x)

\S4method{structures_grob}{ANY}(x) <- value

\S4method{nodes_ggset}{ANY}(x)

\S4method{nodes_ggset}{ANY}(x) <- value

\S4method{nodes_grob}{ANY}(x)

\S4method{nodes_grob}{ANY}(x) <- value

\S4method{ppcp_data}{ANY}(x)

\S4method{ppcp_data}{ANY}(x) <- value

\S4method{ration_data}{ANY}(x)

\S4method{ration_data}{ANY}(x) <- value

\S4method{ggset_annotate}{ANY}(x)

\S4method{ggset_annotate}{ANY}(x) <- value
}
\arguments{
\item{value}{The value for the slot.}
}
\description{
This class store multiple components for visualization.

\code{parent_nebula}: Store data for visualization of
Parent-Nebula.

\code{child_nebulae}: store data for visualization of
Child-Nebulae.

\code{parent_nebula}, \code{parent_nebula<-}: getter and setter
for the \code{parent_nebula} slot of the object.

\code{child_nebulae}, \code{child_nebulae<-}: getter and setter
for the \code{child_nebulae} slot of the object.

\code{igraph}, \code{igraph<-}: getter and setter
for the \code{igraph} slot of the object.

\code{tbl_graph}, \code{tbl_graph<-}: getter and setter
for the \code{tbl_graph} slot of the object.

\code{layout_ggraph}, \code{layout_ggraph<-}: getter and setter
for the \code{layout_ggraph} slot of the object.

\code{grid_layout}, \code{grid_layout<-}: getter and setter
for the \code{grid_layout} slot of the object.

\code{viewports}, \code{viewports<-}: getter and setter
for the \code{viewports} slot of the object.

\code{ggset}, \code{ggset<-}: getter and setter
for the \code{ggset} slot of the object.

\code{panel_viewport}, \code{panel_viewport<-}: getter and setter
for the \code{panel_viewport} slot of the object.

\code{legend_viewport}, \code{legend_viewport<-}: getter and setter
for the \code{legend_viewport} slot of the object.

\code{structures_grob}, \code{structures_grob<-}: getter and setter
for the \code{structures_grob} slot of the object.

\code{nodes_ggset}, \code{nodes_ggset<-}: getter and setter
for the \code{nodes_ggset} slot of the object.

\code{nodes_grob}, \code{nodes_grob<-}: getter and setter
for the \code{nodes_grob} slot of the object.

\code{ppcp_data}, \code{ppcp_data<-}: getter and setter
for the \code{ppcp_data} slot of the object.

\code{ration_data}, \code{ration_data<-}: getter and setter
for the \code{ration_data} slot of the object.

\code{ggset_annotate}, \code{ggset_annotate<-}: getter and setter
for the \code{ggset_annotate} slot of the object.
}
\section{Slots}{

\describe{
\item{\code{parent_nebula}}{\linkS4class{parent_nebula} object.}

\item{\code{child_nebulae}}{\linkS4class{child_nebulae} object.}

\item{\code{igraph}}{"igraph" object or its list. See \code{\link[igraph:graph_from_data_frame]{igraph::graph_from_data_frame()}}.
The slot contains edges and nodes data of Child-Nebulae or Parent-Nebula.
The "igraph" object can be output use \code{\link[igraph:write_graph]{igraph::write_graph()}} as ".graphml" file,
which belong to a network data format that can be operated by other software such as
Cytoscape (\url{https://cytoscape.org/}).}

\item{\code{tbl_graph}}{"tbl_graph" object or its list. See \code{\link[tidygraph:tbl_graph]{tidygraph::as_tbl_graph()}}.
Converted from slot \code{igraph}.}

\item{\code{layout_ggraph}}{"layout_ggraph" object or its list. See \code{\link[ggraph:ggraph]{ggraph::create_layout()}}.
Create from slot \code{tbl_graph}, passed to \code{\link[ggraph:ggraph]{ggraph::ggraph()}} for visualization.}

\item{\code{grid_layout}}{"layout" object. See \code{\link[grid:grid.layout]{grid::grid.layout()}}.
Grid layout for position of each Child-Nebula to visualize.}

\item{\code{viewports}}{list with names. Each element must be "viewport" object.
See \code{\link[grid:viewport]{grid::viewport()}}. Position for each Child-Nebula to visualize.}

\item{\code{panel_viewport}}{"viewport" object. See \code{\link[grid:viewport]{grid::viewport()}}. For visualization,
the position to place overall Child-Nebulae.}

\item{\code{legend_viewport}}{"viewport" object. See \code{\link[grid:viewport]{grid::viewport()}}. For visualization,
the position to place legend.}

\item{\code{ggset}}{\linkS4class{ggset} object or its list with names. Each \linkS4class{ggset} object
can be visualized directly use \code{\link[=call_command]{call_command()}}.}

\item{\code{structures_grob}}{list with names. Each element is a "grob" object.
See \code{\link[grid:grid.grob]{grid::grob()}}. Use \code{\link[grid:grid.draw]{grid::grid.draw()}} to visualize the chemical structure.}

\item{\code{nodes_ggset}}{list of \linkS4class{ggset} object. For drawing each node of 'features'
('features' means the detected peaks while processing LC-MS data)
with annotation. Use \code{\link[=call_command]{call_command()}} to visualize the \linkS4class{ggset}.}

\item{\code{nodes_grob}}{list of "grob" object. Converted from slot \code{nodes_ggset} with slot
\code{structures_grob}. Use \code{\link[grid:grid.draw]{grid::grid.draw()}} to visualize the "grob".}

\item{\code{ppcp_data}}{list with names. Each element is a data.frame. This is an
annotation data of 'features' which would be visualize in nodes border
as a radial bar plot. \code{ppcp_data}, i.e., posterior probability of
classification prediction. See \code{\link[=filter_ppcp]{filter_ppcp()}}.}

\item{\code{ration_data}}{list with names. Each element is a data.frame. This is an
annotation data of 'features' which would be visualize in nodes nucleus as
ring plot. Generally, \code{ration_data} is the statistic data for samples.}

\item{\code{ggset_annotate}}{a list of \linkS4class{ggset} object. The annotated Child-Nebulae
gathered from slot \code{ggset} and slot \code{nodes_grob}.
Use \code{\link[=call_command]{call_command()}} to visualize the \linkS4class{ggset}. Be care, the object
sometimes is too large that need lot of time to loading for visualization.}
}}

\examples{
\dontrun{
new('nebula', ...)
}
\dontrun{
new('parent_nebula', ...)
}
\dontrun{
new('child_nebulae', ...)
}
}
\seealso{
Other nebulae: 
\code{\link{mcnebula-class}}
}
\concept{nebulae}
