% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-plot_msms_mirrors.R
\name{plot_msms_mirrors}
\alias{plot_msms_mirrors}
\title{Draw MS/MS mirror bar plots}
\usage{
plot_msms_mirrors(
  x,
  .features_id,
  fun_modify = modify_set_labs_xy,
  structure_vp = grid::viewport(0.7, 0.3, 0.3, 0.3)
)
}
\arguments{
\item{x}{\linkS4class{mcnebula} object.}

\item{.features_id}{character. The ID of 'features'.}

\item{fun_modify}{function. Used to post modify the \linkS4class{ggset} object before
visualization. See \link{fun_modify}.}

\item{structure_vp}{'viewport' object. Created by \code{\link[grid:viewport]{grid::viewport()}}. The 'viewport'
to draw chemical structures in sub-panel.}
}
\description{
Draw MS/MS spectra as mirror bar plots using 'ggplot2' with \code{\link[=facet_wrap]{facet_wrap()}}.
The sub-panel of each 'features' would be found by \code{\link[grid:grid.grep]{grid::grid.grep()}}...
Then chemical structures would be drawn into sub-panel.
}
\seealso{
\code{\link[=facet_wrap]{facet_wrap()}}, \link[grid:grid.force]{grid::grid.force}, \code{\link[grid:grid.grep]{grid::grid.grep()}}...
}
