% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-project.R
\docType{class}
\name{project-class}
\alias{project-class}
\alias{.project}
\alias{project}
\alias{project_version,ANY-method}
\alias{project_version}
\alias{project_version<-,ANY-method}
\alias{project_version<-}
\alias{project_path,ANY-method}
\alias{project_path}
\alias{project_path<-,ANY-method}
\alias{project_path<-}
\alias{file_name,ANY-method}
\alias{file_name}
\alias{file_api,ANY-method}
\alias{file_api}
\alias{attribute_name,ANY-method}
\alias{attribute_name}
\alias{metadata,ANY-method}
\alias{project_metadata}
\alias{methods_read,ANY-method}
\alias{methods_read}
\alias{methods_format,ANY-method}
\alias{methods_format}
\alias{methods_match,ANY-method}
\alias{methods_match}
\alias{match.candidates_id,ANY-method}
\alias{match.candidates_id}
\alias{match.features_id,ANY-method}
\alias{match.features_id}
\alias{get_upper_dir_subscript,ANY,character,missing-method}
\alias{get_upper_dir_subscript}
\title{Collection of Interface for extracting data from raw directory}
\usage{
\S4method{project_version}{ANY}(x)

\S4method{project_version}{ANY}(x) <- value

\S4method{project_path}{ANY}(x)

\S4method{project_path}{ANY}(x) <- value

\S4method{file_name}{ANY}(x)

\S4method{file_api}{ANY}(x)

\S4method{attribute_name}{ANY}(x)

\S4method{metadata}{ANY}(x)

\S4method{methods_read}{ANY}(x)

\S4method{methods_format}{ANY}(x)

\S4method{methods_match}{ANY}(x)

\S4method{match.candidates_id}{ANY}(x)

\S4method{match.features_id}{ANY}(x)

\S4method{get_upper_dir_subscript}{ANY,character,missing}(x, subscript)
}
\arguments{
\item{x}{...}

\item{value}{The value for the slot.}

\item{subscript}{...}
}
\description{
...

\code{project_version}, \code{project_version<-}: getter and setter
for the \code{project_version} slot of the object.

\code{project_path}, \code{project_path<-}: getter and setter
for the \code{project_path} slot of the object.

\code{file_name}, \code{file_api}, \code{attribute_name}:
fast channel to obtain
the downstream slot. e.g., getter
for the \code{file_name} slot in sub-object
of \code{project_conformation} slot of the object. Equals:
\itemize{
\item \code{file_name(project_conformation(object))}
\item \code{file_name(object)}.
}

\code{project_metadata}: fast channel to obtain
the downstream slot, getter
for the \code{project_metadata} slot in sub-object
of \code{project_metadata} slot of the object. Equals:
\itemize{
\item \code{project_metadata(project_metadata(object))}
\item \code{project_metadata(object)}.
}

\code{methods_read}, \code{methods_format}, \code{methods_match}:
fast channel to obtain
the downstream slot. e.g., getter
for the \code{methods_read} slot in sub-object
of \code{project_api} slot of the object. Equals:
\itemize{
\item \code{methods_read(project_api(object))}
\item \code{methods_read(object)}.
}

\code{match.candidates_id}, \code{match.features_id}:
fast channel to obtain
data (mostly 'tbl' or 'data.frame') inside the downstream slot ('list'), getter
for the data named \code{match.candidates_id} in
\code{methods_match} slot (a 'list') in sub-object
of \code{project_api} slot of the object. Equals:
\itemize{
\item \code{methods_match(project_api(object))$match.candidates_id}
\item \code{match.candidates_id(object)}.
}

\code{get_upper_dir_subscript}: ...
}
\section{Slots}{

\describe{
\item{\code{project_version}}{character.}

\item{\code{project_path}}{...}

\item{\code{project_conformation}}{...}

\item{\code{project_metadata}}{...}

\item{\code{project_api}}{...}

\item{\code{project_dataset}}{...}
}}

\examples{
\dontrun{
new('project', ...)
}
\dontrun{
get_upper_dir_subscript(...)
}
}
\seealso{
Other projects: 
\code{\link{project_api-class}},
\code{\link{project_conformation-class}},
\code{\link{project_dataset-class}},
\code{\link{project_metadata-class}}
}
\concept{projects}
