% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-project_api.R
\docType{class}
\name{project_api-class}
\alias{project_api-class}
\alias{.project_api}
\alias{project_api}
\alias{show,project_api-method}
\alias{show}
\alias{project_api,ANY-method}
\alias{project_api<-,ANY-method}
\alias{project_api<-}
\alias{methods_read,project_api-method}
\alias{methods_read}
\alias{methods_read<-,project_api-method}
\alias{methods_read<-}
\alias{methods_format,project_api-method}
\alias{methods_format}
\alias{methods_format<-,project_api-method}
\alias{methods_format<-}
\alias{methods_match,project_api-method}
\alias{methods_match}
\alias{methods_match<-,project_api-method}
\alias{methods_match<-}
\title{Function set for extracting data}
\usage{
\S4method{show}{project_api}(object)

\S4method{project_api}{ANY}(x)

\S4method{project_api}{ANY}(x) <- value

\S4method{methods_read}{project_api}(x)

\S4method{methods_read}{project_api}(x) <- value

\S4method{methods_format}{project_api}(x)

\S4method{methods_format}{project_api}(x) <- value

\S4method{methods_match}{project_api}(x)

\S4method{methods_match}{project_api}(x) <- value
}
\arguments{
\item{value}{The value for the slot.}
}
\description{
This is a class object used to store various functions for extracting and formatting data.
See \linkS4class{project} for joint application with other related classes.

\code{project_api}, \code{project_api<-}: getter and setter
for the \code{project_api} slot of the object.

\code{methods_read}, \code{methods_read<-}: getter and setter
for the \code{methods_read} slot of the object.

\code{methods_format}, \code{methods_format<-}: getter and setter
for the \code{methods_format} slot of the object.

\code{methods_match}, \code{methods_match<-}: getter and setter
for the \code{methods_match} slot of the object.
}
\section{Slots}{

\describe{
\item{\code{methods_read}}{list. Store a list of functions for reading data.
The list with the names: "read" + "subscript". e.g., "read.f3_fingerid".}

\item{\code{methods_format}}{function. The function is used to format the data
(e.g., rename the column names; convert the columns of character type into numeric).}

\item{\code{methods_match}}{list. Store a list of functions for matching and extracting string.}
}}

\note{
The class is not for normal use of the package.
}
\seealso{
Other projects: 
\code{\link{project-class}},
\code{\link{project_conformation-class}},
\code{\link{project_dataset-class}},
\code{\link{project_metadata-class}}
}
\concept{projects}
