% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-project_conformation.R
\docType{class}
\name{project_conformation-class}
\alias{project_conformation-class}
\alias{.project_conformation}
\alias{project_conformation}
\alias{show,project_conformation-method}
\alias{show}
\alias{project_conformation,ANY-method}
\alias{project_conformation<-,ANY-method}
\alias{project_conformation<-}
\alias{file_name,project_conformation-method}
\alias{file_name}
\alias{file_name<-,project_conformation-method}
\alias{file_name<-}
\alias{file_api,project_conformation-method}
\alias{file_api}
\alias{file_api<-,project_conformation-method}
\alias{file_api<-}
\alias{attribute_name,project_conformation-method}
\alias{attribute_name}
\alias{attribute_name<-,project_conformation-method}
\alias{attribute_name<-}
\title{Clarify the name, path and attribute name of files
in the project (directory)}
\usage{
\S4method{show}{project_conformation}(object)

\S4method{project_conformation}{ANY}(x)

\S4method{project_conformation}{ANY}(x) <- value

\S4method{file_name}{project_conformation}(x)

\S4method{file_name}{project_conformation}(x) <- value

\S4method{file_api}{project_conformation}(x)

\S4method{file_api}{project_conformation}(x) <- value

\S4method{attribute_name}{project_conformation}(x)

\S4method{attribute_name}{project_conformation}(x) <- value
}
\arguments{
\item{value}{The value for the slot.}
}
\description{
...

\code{project_conformation}, \code{project_conformation<-}: getter and setter
for the \code{project_conformation} slot of the object.

\code{file_name}, \code{file_name<-}: getter and setter
for the \code{file_name} slot of the object.

\code{file_api}, \code{file_api<-}: getter and setter
for the \code{file_api} slot of the object.

\code{attribute_name}, \code{attribute_name<-}: getter and setter
for the \code{attribute_name} slot of the object.
}
\section{Slots}{

\describe{
\item{\code{file_name}}{...}

\item{\code{file_api}}{...}

\item{\code{attribute_name}}{...}
}}

\examples{
\dontrun{
new('project_conformation', ...)
}
}
\seealso{
Other projects: 
\code{\link{project-class}},
\code{\link{project_api-class}},
\code{\link{project_dataset-class}},
\code{\link{project_metadata-class}}
}
\concept{projects}
