% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-project_metadata.R
\docType{class}
\name{project_metadata-class}
\alias{project_metadata-class}
\alias{.project_metadata}
\alias{project_metadata}
\alias{show,project_metadata-method}
\alias{show}
\alias{project_metadata,ANY-method}
\alias{project_metadata<-,ANY-method}
\alias{project_metadata<-}
\alias{latest,project_metadata,ANY,ANY-method}
\alias{latest}
\alias{metadata,project_metadata-method}
\alias{metadata}
\alias{metadata<-,project_metadata-method}
\alias{metadata<-}
\alias{add_dataset,project_metadata,list-method}
\alias{add_dataset}
\alias{extract_metadata,ANY,character-method}
\alias{extract_metadata}
\alias{get_metadata,ANY,character,ANY,ANY,ANY-method}
\alias{get_metadata}
\alias{get_metadata,missing,character,project_metadata,project_conformation,character-method}
\title{Metadata of files}
\usage{
\S4method{show}{project_metadata}(object)

\S4method{project_metadata}{ANY}(x)

\S4method{project_metadata}{ANY}(x) <- value

\S4method{latest}{project_metadata,ANY,ANY}(x)

\S4method{metadata}{project_metadata}(x)

\S4method{metadata}{project_metadata}(x) <- value

\S4method{add_dataset}{project_metadata,list}(x, list)

\S4method{extract_metadata}{ANY,character}(x, subscript)

\S4method{get_metadata}{ANY,character,ANY,ANY,ANY}(x, subscript)

\S4method{get_metadata}{missing,character,project_metadata,project_conformation,character}(subscript, project_metadata, project_conformation, path)
}
\arguments{
\item{value}{The value for the slot.}

\item{list}{a list (with names) of metadata (data.frame) with names.}

\item{subscript}{see \linkS4class{subscript}.}

\item{project_metadata}{\linkS4class{project_metadata} object.
Used by \code{get_metadata()}. If 'missing', the slot \code{project_metadata} inside
the object will be used.}

\item{project_conformation}{\linkS4class{project_conformation} object.
Used by \code{get_metadata()}. If 'missing', the slot \code{project_conformation} inside
the object will be used.}

\item{path}{character. The path of the project directory (generally, SIRIUS project).
If 'missing', the slot \code{project_path} inside the object will be used.}
}
\description{
This is a class object used to store metadata of files.
See \linkS4class{project} for joint application with other related classes.

\code{project_metadata}, \code{project_metadata<-}: getter and setter
for the \code{project_metadata} slot of the object.

\code{latest}: get the first data in \code{metadata} slot and
format as "tbl".

\code{metadata}, \code{metadata<-}: getter and setter
for the \code{metadata} slot of the object.

\code{add_dataset}: add the list into slot \code{metadata}.

\code{extract_metadata}: use "subscript" to extract metadata from an
object with slot \code{project_metadata},
and then return it as a new \code{project_metadata}.

\code{get_metadata}: for an object with slot of \code{project_metadata},
get the metadata of files of specified "subscript", then return the object.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{a list with names of \linkS4class{subscript}.
Each element of the list is a data.frame.}
}}

\note{
The class is not for normal use of the package.
}
\seealso{
Other projects: 
\code{\link{project-class}},
\code{\link{project_api-class}},
\code{\link{project_conformation-class}},
\code{\link{project_dataset-class}}

Other latests: 
\code{\link{mcn_dataset-class}},
\code{\link{mcnebula-class}},
\code{\link{msframe-class}},
\code{\link{project_dataset-class}}
}
\concept{latests}
\concept{projects}
