% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-report.R
\docType{class}
\name{report-class}
\alias{report-class}
\alias{.report}
\alias{report}
\alias{show_layers,report-method}
\alias{show_layers}
\alias{yaml,ANY-method}
\alias{yaml}
\alias{yaml<-,ANY-method}
\alias{yaml<-}
\alias{new_report,character-method}
\alias{new_report}
\alias{new_report,missing-method}
\alias{call_command,report-method}
\alias{call_command}
\alias{search_heading}
\title{Creating a formatted report}
\usage{
\S4method{show_layers}{report}(x)

\S4method{yaml}{ANY}(x)

\S4method{yaml}{ANY}(x) <- value

\S4method{new_report}{character}(..., yaml)

\S4method{new_report}{missing}(..., yaml)

\S4method{call_command}{report}(x)

search_heading(x, pattern, level = NULL)
}
\arguments{
\item{x}{\linkS4class{report} object.}

\item{value}{The value for the slot.}

\item{...}{An arbitrary number of \linkS4class{heading},
\linkS4class{section} or \linkS4class{code_block} in sequence.
Specially, \code{NULL} can be passed herein, but would be ignored.}

\item{yaml}{character. Passed to .Rmd for setting format of documentation.}

\item{pattern}{character(1). For Regex match. Allowed Perl expression.}

\item{level}{numeric.
in slot \code{layers}.}
}
\description{
The report module can create output report quickly for
and not just for the mcnebula2 workflow.
The report system is primarily a class object that manages text and code blocks.
Heading or paragraphs or code blocks were treated as individual report units
and deposited sequentially in "layers".
The report system provides methods to exhibit, modify these layers.
Reports can be exported as ".Rmd" text files, and subsequently,
users can call \code{\link[rmarkdown:render]{rmarkdown::render()}} for output formatted documents.

\code{show_layers}: show \code{layers} slots in a pretty
and readable style.

\code{yaml}, \code{yaml<-}: getter and setter
for the \code{yaml} slot of the object.

\code{new_report}: Create a \linkS4class{report} object.

\code{new_report()}: get the default parameters for the method \code{new_report}.

\code{new_report(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{new_report}.

\code{call_command}: Format 'report' object as character, which can be output
by \code{writeLines()} function as '.Rmd' file and than use \code{rmarkdown::render} output
as pdf, html, or other format files.

\code{search_heading}: Regex match for \linkS4class{heading} object
}
\section{Slots}{

\describe{
\item{\code{yaml}}{character. Metadata passed to .Rmd for setting format of documentation.
See \url{https://bookdown.org/yihui/rmarkdown/compile.html} for details.}

\item{\code{layers}}{list. Element in list must be \linkS4class{section},
\linkS4class{heading} or \linkS4class{code_block}.}
}}

\examples{
\dontrun{
  s1 <- new_heading("Title 1", 1)
  
  s1.5 <- new_section2(
    c("..."), NULL
  )
  
  s2 <- new_section2(
    c("This is sentence 1.",
      "This is sentence 2.",
      "This is sentence 3."),
    rblock({
      df <- data.frame(x = 1:10, y = 1:10)
      df <- dplyr::mutate(df,
        group = rep(paste0("p", 1:3), c(3, 3, 4))
      )
    })
  )
  
  s3 <- new_heading("Title 2", 1)
  
  s4 <- new_section2(
    c("This is a paragraph..."),
    rblock({
      p <- ggplot(df) +
        geom_point(aes(x = x, y = y, fill = group))
      ggsave(f <- paste0(tempdir(), "/test.pdf"), p)
    })
  )
  
  s5 <- include_figure(f, "name", "Caption: ...")
  
  s6 <- rblock({
    print(1:100)
  }, args = list(echo = F, eval = F))
  
  s7 <- c("... paragraph ...")
  
  sections <- gather_sections()
  report <- do.call(new_report, sections)
  render_report(report, paste0(tempdir(), "/report.Rmd"), T)
  
  ####################
  ##### Another example
  ####################
  
  h1 <- new_heading("heading, level 1", 1)
  
  sec1 <- new_section(
    "sub-heading", 2,
    "This is a description.",
    new_code_block(codes = "seq <- lapply(1:10, cat)")
  )
  
  h2 <- new_heading("heading 2, level 1", 1)
  
  fig_block <- new_code_block_figure("plot", "this is a caption",
    codes = "df <- data.frame(x = 1:10, y = 1:10)
      p <- ggplot(df) +
        geom_point(aes(x = x, y = y))
      p"
  )
  sec2 <- new_section(
    "sub-heading2", 2,
    paste0(
      "This is a description. ",
      "See Figure ", get_ref(fig_block), "."
    ),
    fig_block
  )
  
  a_data <- dplyr::storms[1:15, 1:10]
  table_block <- include_table(a_data, "table1", "This is a caption")
  
  sec3 <- new_section(
    NULL, ,
    paste0("See Table ", get_ref(table_block, "tab"), "."),
    NULL
  )
  
  tmp_p <- paste0(tempdir(), "/test.pdf")
  pdf(tmp_p)
  plot(1:10)
  dev.off()
  fig_block_2 <- include_figure(tmp_p, "plot2", "this is a caption")
  sec4 <- history_rblock(, "^tmp_p <- ", "^fig_block_2")
  sec4
  
  ## gather
  yaml <- "title: 'title'\noutput:\n  bookdown::pdf_document2"
  report <- new_report(
    h1, sec1, h2, sec2,
    table_block, sec3,
    fig_block_2, sec4,
    yaml = yaml
  )
  report
  
  ## output
  tmp <- paste0(tempdir(), "/tmp_output.Rmd")
  render_report(report, tmp)
  rmarkdown::render(tmp)
  file.exists(sub("Rmd$", "pdf", tmp))
}
}
\seealso{
\code{\link[=writeLines]{writeLines()}}, \code{\link[rmarkdown:render]{rmarkdown::render()}}...

Other layerSets: 
\code{\link{ggset-class}},
\code{\link{layerSet-class}}

Other call_commands: 
\code{\link{code_block-class}},
\code{\link{command-class}},
\code{\link{ggset-class}},
\code{\link{section-class}}
}
\concept{call_commands}
\concept{layerSets}
