% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-section.R
\docType{class}
\name{section-class}
\alias{section-class}
\alias{.section}
\alias{section}
\alias{heading-class}
\alias{.heading}
\alias{heading}
\alias{heading,ANY-method}
\alias{heading<-,ANY-method}
\alias{heading<-}
\alias{level,heading-method}
\alias{level}
\alias{level<-,heading-method}
\alias{level<-}
\alias{new_heading,character,numeric-method}
\alias{new_heading}
\alias{call_command,heading-method}
\alias{call_command}
\alias{paragraph,section-method}
\alias{paragraph}
\alias{paragraph<-,section-method}
\alias{paragraph<-}
\alias{new_section,missing,missing,missing,missing-method}
\alias{new_section,ANY,ANY,ANY,ANY-method}
\alias{new_section,character,numeric,character,maybe_code_block-method}
\alias{new_section}
\alias{new_section,NULL,numeric,character,maybe_code_block-method}
\alias{new_section2}
\alias{call_command,section-method}
\alias{call_command,NULL-method}
\title{Basic cells in the report}
\usage{
\S4method{heading}{ANY}(x)

\S4method{heading}{ANY}(x) <- value

\S4method{level}{heading}(x)

\S4method{level}{heading}(x) <- value

\S4method{new_heading}{character,numeric}(heading, level)

\S4method{call_command}{heading}(x)

\S4method{paragraph}{section}(x)

\S4method{paragraph}{section}(x) <- value

\S4method{new_section}{missing,missing,missing,missing}(heading, level, paragraph, code_block)

\S4method{new_section}{ANY,ANY,ANY,ANY}(heading, level, paragraph, code_block)

\S4method{new_section}{character,numeric,character,maybe_code_block}(heading, level, paragraph, code_block)

\S4method{new_section}{`NULL`,numeric,character,maybe_code_block}(heading, level, paragraph, code_block)

new_section2(paragraph, code_block)

\S4method{call_command}{section}(x)

\S4method{call_command}{`NULL`}(x)
}
\arguments{
\item{value}{The value for the slot.}

\item{heading}{character(1). For slot \code{.Data}.}

\item{level}{numeric(1). For slot \code{level}.}

\item{paragraph}{character. Text for description.}

\item{code_block}{\linkS4class{code_block} object.}
}
\description{
A class object consist of \linkS4class{heading}, paragraph (character),
and \linkS4class{code_block}. These \linkS4class{section} belong to basic cells of report.

This is a class object used to clarify the heading and its hierarchy.

\code{heading}, \code{heading<-}: getter and setter
for the \code{heading} slot of the object.

\code{level}, \code{level<-}: getter and setter
for the \code{level} slot of the object.

\code{new_heading}: create \linkS4class{heading} object.

\code{call_command}: Format 'heading' object as character.

\code{paragraph}, \code{paragraph<-}: getter and setter
for the \code{paragraph} slot of the object.

\code{new_section()}: get the default parameters for
the method \code{new_section}.

\code{new_section(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{new_section}.

\code{new_section}: create \linkS4class{section} object.

\code{new_section2}: Identical to \code{new_section(NULL, , ...)}

\code{call_command}: Format 'section' object as character.
}
\section{Slots}{

\describe{
\item{\code{heading}}{\linkS4class{heading} object.}

\item{\code{paragraph}}{character. Text for description.}

\item{\code{code_block}}{\linkS4class{code_block} object.}

\item{\code{.Data}}{character(1). Text of heading.}

\item{\code{level}}{numeric. Level of heading.}
}}

\examples{
\dontrun{
  ## ------------------------------------- 
  ## heading
  new_heading("this is a heading", 2)
  
  ## ------------------------------------- 
  ## section
  ## example 1
  para <- "This is a paragraph stating"
  section <- new_section("this is a heading", 2, para)
  ## see results
  section
  call_command(section)
  writeLines(call_command(section))
  
  ## example 2
  para <- "This is a paragraph stating"
  section <- new_section(NULL, , para, NULL)
  
  ## example 3
  para <- "This is a paragraph stating"
  block <- new_code_block(codes = "df <- data.frame(x = 1:10)")
  section <- new_section("heading", 2, para, block)
  section
  
  ## example 4
  codes <- "df <- data.frame(x = 1:10, y = 1:10)
    p <- ggplot(df) +
      geom_point(aes(x = x, y = y))
    p"
  fig_block <- new_code_block_figure("plot", "this is caption", codes = codes)
  para <- paste0("This is a paragraph describing the picture. ",
                 "See Figure ", get_ref(fig_block), ".")
  section <- new_section("heading", 2, para, fig_block)
  section
  ## output
  tmp <- paste0(tempdir(), "/tmp_output.Rmd")
  writeLines(call_command(section), tmp)
  rmarkdown::render(tmp, output_format = "bookdown::pdf_document2")
  file.exists(sub("Rmd$", "pdf", tmp))
  ## see [report-class] object: 
  ## A complete output report, including multiple 'section'.
  
  ## defalt parameters
  new_section()
}
}
\seealso{
Other call_commands: 
\code{\link{code_block-class}},
\code{\link{command-class}},
\code{\link{ggset-class}},
\code{\link{report-class}}
}
\concept{call_commands}
