% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-activate_nebulae.R
\name{set_nodes_color-methods}
\alias{set_nodes_color-methods}
\alias{set_nodes_color}
\alias{set_nodes_color,mcnebula,character,data.frame,missing-method}
\alias{set_nodes_color,mcnebula,character,missing,missing-method}
\alias{set_nodes_color,mcnebula,missing,missing,logical-method}
\title{...}
\usage{
\S4method{set_nodes_color}{mcnebula,character,data.frame,missing}(x, attribute, extra_data)

\S4method{set_nodes_color}{mcnebula,character,missing,missing}(x, attribute)

\S4method{set_nodes_color}{mcnebula,missing,missing,logical}(x, use_tracer)
}
\arguments{
\item{x}{...}

\item{attribute}{...}

\item{extra_data}{...}

\item{use_tracer}{...}
}
\value{
...
}
\description{
...

...
}
\details{
...

...
}
\examples{
\dontrun{
set_nodes_color(...)
}
}
\seealso{
\code{\link[=activate_nebulae]{activate_nebulae()}}, \code{\link[=visualize]{visualize()}}...
}
