% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-activate_nebulae.R
\name{set_nodes_color,mcnebula,character-method}
\alias{set_nodes_color,mcnebula,character-method}
\alias{set_nodes_color}
\alias{set_nodes_color,mcnebula,missing-method}
\title{...}
\usage{
\S4method{set_nodes_color}{mcnebula,character}(x, attribute)

\S4method{set_nodes_color}{mcnebula,missing}(x)
}
\arguments{
\item{param}{...}
}
\value{
...
}
\description{
...

\code{set_nodes_color()}: list the recommended attribute for the method
\code{set_nodes_color}.
}
\details{
...
}
\examples{
\dontrun{
set_nodes_color(...)
}
}
\seealso{
\code{\link[=activate_nebulae]{activate_nebulae()}}, \code{\link[=visualize]{visualize()}}...
}
