% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraMethods-draw_nodes.R
\name{set_ppcp_data,mcnebula,character-method}
\alias{set_ppcp_data,mcnebula,character-method}
\alias{set_ppcp_data,missing,missing-method}
\alias{set_ppcp_data,mcnebula,ANY-method}
\title{...}
\usage{
\S4method{set_ppcp_data}{mcnebula,character}(x, classes)

\S4method{set_ppcp_data}{missing,missing}()

\S4method{set_ppcp_data}{mcnebula,ANY}(x, classes)
}
\arguments{
\item{x}{...}

\item{classes}{...}
}
\description{
...

\code{set_ppcp_data()}: get the function for generating
default parameters for the method
\code{set_ppcp_data}.

\code{set_ppcp_data(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{set_ppcp_data}.
}
\examples{
\dontrun{
set_ppcp_data(...)
}
}
