% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraMethods-draw_nodes.R
\name{set_ration_data,mcnebula,logical-method}
\alias{set_ration_data,mcnebula,logical-method}
\alias{set_ration_data,missing,missing-method}
\alias{set_ration_data,mcnebula,ANY-method}
\title{...}
\usage{
\S4method{set_ration_data}{mcnebula,logical}(x, mean)

\S4method{set_ration_data}{missing,missing}()

\S4method{set_ration_data}{mcnebula,ANY}(x, mean)
}
\arguments{
\item{x}{...}

\item{mean}{...}
}
\description{
...

\code{set_ration_data()}: get the default parameters for the method
\code{set_ration_data}.

\code{set_ration_data(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{set_ration_data}.
}
\examples{
\dontrun{
set_ration_data(...)
}
}
