% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-create_nebula_index.R
\name{set_tracer-methods}
\alias{set_tracer-methods}
\alias{set_tracer}
\alias{set_tracer,missing,missing,missing,missing-method}
\alias{set_tracer,mcnebula,ANY,ANY,ANY-method}
\alias{set_tracer,mcnebula,character,character,character-method}
\title{...}
\usage{
\S4method{set_tracer}{missing,missing,missing,missing}()

\S4method{set_tracer}{mcnebula,ANY,ANY,ANY}(x, .features_id, colors, rest)

\S4method{set_tracer}{mcnebula,character,character,character}(x, .features_id, colors, rest)
}
\arguments{
\item{x}{...}

\item{.features_id}{...}

\item{colors}{...}

\item{rest}{...}
}
\value{
...
}
\description{
...

\code{set_tracer()}: get the function for generating
default parameters for the method
\code{set_tracer}.

\code{set_tracer(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{set_tracer}.

...
}
\details{
...

...
}
\examples{
\dontrun{
set_tracer(...)
}
}
\seealso{
\code{\link[=create_nebula_index]{create_nebula_index()}}
}
