% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.font}
\alias{.font}
\title{Set font for visualization of MCnebula2}
\format{
An object of class \code{character} of length 1.
}
\usage{
.font
}
\arguments{
\item{font}{character(1). Such as 'Times'. If you output the
visualization for pdf, use \code{grDevices::pdfFonts()} to checkout
the available fonts; else, you might need help with package \code{extrafont}.}
}
\description{
\bold{Note that} your R harbours the font you set.
}
\keyword{datasets}
