% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VIRTUAL_slots.R
\docType{class}
\name{layerSet-class}
\alias{layerSet-class}
\alias{VIRTUAL_layerSet}
\alias{layerSet}
\alias{layers,layerSet-method}
\alias{layers}
\alias{layers<-,layerSet-method}
\alias{layers<-}
\title{Share slots and methods for classes inherite from VIRTUAL_layerSet}
\usage{
\S4method{layers}{layerSet}(x)

\S4method{layers}{layerSet}(x) <- value
}
\arguments{
\item{value}{The value for the slot.}
}
\description{
...

\code{layers}, \code{layers<-}: getter and setter
for the \code{layers} slot of the object.
}
\section{Slots}{

\describe{
\item{\code{layers}}{...}
}}

\seealso{
Other layerSets: 
\code{\link{report-class}}
}
\concept{layerSets}
