% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-visualize.R
\name{visualize-methods}
\alias{visualize-methods}
\alias{visualize}
\alias{visualize,mcnebula,missing,ANY,missing-method}
\alias{visualize,missing,missing,missing,missing-method}
\alias{visualize,mcnebula,ANY,ANY,ANY-method}
\alias{visualize,mcnebula,character,function,missing-method}
\alias{visualize,mcnebula,numeric,function,missing-method}
\alias{visualize,mcnebula,numeric_or_character,function,logical-method}
\alias{visualize_all,missing,missing,missing,missing-method}
\alias{visualize_all,mcnebula,ANY,ANY,ANY-method}
\alias{visualize_all,mcnebula,logical,function,logical-method}
\title{Visualize Nebulae in R graphic device}
\usage{
\S4method{visualize}{mcnebula,missing,ANY,missing}(x, fun_modify)

\S4method{visualize}{missing,missing,missing,missing}()

\S4method{visualize}{mcnebula,ANY,ANY,ANY}(x, item, fun_modify, annotate)

\S4method{visualize}{mcnebula,character,`function`,missing}(x, item, fun_modify)

\S4method{visualize}{mcnebula,numeric,`function`,missing}(x, item, fun_modify)

\S4method{visualize}{mcnebula,numeric_or_character,`function`,logical}(x, item, fun_modify, annotate)

\S4method{visualize_all}{missing,missing,missing,missing}()

\S4method{visualize_all}{mcnebula,ANY,ANY,ANY}(x, newpage, fun_modify, legend_hierarchy)

\S4method{visualize_all}{mcnebula,logical,`function`,logical}(x, newpage, fun_modify, legend_hierarchy)
}
\arguments{
\item{x}{\linkS4class{mcnebula} object.}

\item{fun_modify}{function. Used to post modify the \linkS4class{ggset} object before
visualization. See \link{fun_modify}.}

\item{item}{character(1) or numeric(1). If \code{character}, the value should be
a name of chemical class in 'nebula_index' data. Its Nebulae has been activated
via \code{\link[=activate_nebulae]{activate_nebulae()}}. If \code{numeric}, the value should be the sequence of
Nebulae... Use \code{visualize(object)} to get the optional value.}

\item{annotate}{logical. If \code{TRUE}, visualize the Nebula with the annotation.
Only available \code{\link[=annotate_nebula]{annotate_nebula()}} has been run for the Nebula.}

\item{newpage}{logical. If \code{TRUE}, use \code{\link[grid:grid.newpage]{grid::grid.newpage()}} before visualization.}

\item{legend_hierarchy}{logical. If \code{TRUE}, visualize the legend of chemical hierarchy.}
}
\description{
Methods used for visualization.
Show chemical Nebulae (either Parent-Nebula or Child-Nebulae) in R graphic device.
Run after \code{\link[=activate_nebulae]{activate_nebulae()}}

\code{visualize(x)}: get a 'tbl' about Child-Nebulae candidates
for \code{visualize} methods to visualize.

\code{visualize()}: get the default parameters for the method
\code{visualize}.

\code{visualize(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{visualize}.

\code{visualize_all()}: get the default parameters for the method
\code{visualize_all}.

\code{visualize_all(x, ...)}: use the default parameters whatever 'missing'
while performing the method \code{visualize_all}.

\code{visualize_all}: visualize overall Child-Nebulae into R graphic device.
}
\examples{
\dontrun{
visualize(...)
}
\dontrun{
visualize_all(...)
}
}
